% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{alazygatr_packages}
\alias{alazygatr_packages}
\title{Install alazygatr packages}
\usage{
alazygatr_packages()
}
\value{
None
}
\description{
Checks for the presence of packages required for running alazygatr.
If any of these packages are not already installed, it will automatically install them.
}
\examples{
\dontrun{alazygatr_packages()}
}
