% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{allK_plot_helper}
\alias{allK_plot_helper}
\title{Helper function for all K plotting}
\usage{
allK_plot_helper(K, krig_admix, coords = NULL, col, col_breaks, ...)
}
\arguments{
\item{K}{K value}

\item{krig_admix}{RasterStack returned by \link[landgen]{tess_krig}}

\item{coords}{dataframe with x and y coordinates for plotting (optional)}

\item{col}{single color code}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to base plot.}
}
\description{
Helper function for all K plotting
}
