% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{bestK}
\alias{bestK}
\title{Best K Selection based on cross entropy}
\usage{
bestK(tess3_obj, Kvals)
}
\arguments{
\item{tess3_obj}{list produced by \code{\link{tess3}}}

\item{Kvals}{vector of K values for testing}
}
\description{
Best K Selection based on cross entropy
}
\note{
(source: https://chazhyseni.github.io/NALgen/post/determining_bestk/)
}
\seealso{
Other TESS functions: 
\code{\link{allK_plot_helper}()},
\code{\link{ggbarplot_helper}()},
\code{\link{krig_legend}()},
\code{\link{pops_helper}()},
\code{\link{tess_barplot}()},
\code{\link{tess_col_default}()},
\code{\link{tess_do_everything}()},
\code{\link{tess_ggbarplot}()},
\code{\link{tess_ggplot}()},
\code{\link{tess_krig}()},
\code{\link{tess_ktest}()},
\code{\link{tess_plot_allK}()}
}
\concept{TESS functions}
