% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CA_env}
\alias{CA_env}
\title{Example environmental data, calculated by performing a raster PCA on 18 bioclimatic variables for state of California}
\format{
A RasterStack object with three environmental PCs
}
\usage{
CA_env
}
\description{
Example environmental data, calculated by performing a raster PCA on 18 bioclimatic variables for state of California
}
\keyword{datasets}
