% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vars.R
\name{check_dists}
\alias{check_dists}
\title{Check geographic and environmental distances for collinearity}
\usage{
check_dists(envlayers, coords, type = "Euclidean", lyr = NULL, sig = 0.05)
}
\arguments{
\item{envlayers}{SpatRaster or Raster* object}

\item{coords}{dataframe with x and y sample coordinates}

\item{type}{the type of geographic distance to be calculated; options are "Euclidean" for direct distance, "topographic" for topographic distances, and "resistance" for resistance distances}

\item{lyr}{DEM raster for calculating topographic distances or resistance raster for calculating resistance distances}

\item{sig}{significance threshold for Mantel test}
}
\value{
a list with (1) a dataframe of significantly correlated variables, (2) a matrix of p-values, (3) a matrix of Mantel's r
}
\description{
Check geographic and environmental distances for collinearity
}
