% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vars.R
\name{check_env}
\alias{check_env}
\title{Check environmental layers for collinearity}
\usage{
check_env(envlayers, threshold = 0.5)
}
\arguments{
\item{envlayers}{A RasterStack of data layers}

\item{threshold}{The cutoff correlation coefficient for flagging variables as collinear (numeric)}
}
\value{
A matrix of correlation coefficients
}
\description{
Check environmental layers for collinearity
}
