% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_raster.R
\name{coords_to_raster}
\alias{coords_to_raster}
\title{Create a raster from coordinates}
\usage{
coords_to_raster(
  coords,
  buff = 0,
  res = NULL,
  agg = NULL,
  disagg = NULL,
  plot = FALSE
)
}
\arguments{
\item{coords}{x and y coordinates (two columns, the first should be x and the second should be y)}

\item{buff}{buffer to add to edge of raster (default = 0)}

\item{res}{desired resolution of raster (default = NULL). Can be a single value for square cells or a vector with two values representing x and y resolutions.}

\item{agg}{aggregation factor to apply to raster (default = NULL)}

\item{disagg}{disaggregation factor to apply to raster (default = NULL)}

\item{plot}{whether to plot resulting raster with coords (default = FALSE)}
}
\value{
RasterLayer
}
\description{
Create a raster from coordinates
}
\examples{
load_mini_ex()
coords_to_raster(mini_coords, buffer = 1, plot = TRUE)
}
