% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_convert.R
\name{coords_to_sf}
\alias{coords_to_sf}
\title{Convert from matrix, data frame, or sf to sf}
\usage{
coords_to_sf(coords)
}
\arguments{
\item{coords}{sf object, data frame, or matrix representing coordinates}
}
\value{
converted coords in sf format
}
\description{
Convert from matrix, data frame, or sf to sf
}
