% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_convert.R
\name{coords_to_sp}
\alias{coords_to_sp}
\title{Convert from matrix, data frame, or sf to formatted sp}
\usage{
coords_to_sp(coords)
}
\arguments{
\item{coords}{sf object, data frame, or matrix representing coordinates}
}
\value{
converted coords in sp format
}
\description{
Convert from matrix, data frame, or sf to formatted sp
}
