% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{data_processing_packages}
\alias{data_processing_packages}
\title{Install data processing packages}
\usage{
data_processing_packages()
}
\value{
None
}
\description{
Checks for the presence of packages required for genetic data processing.
If any of these packages are not already installed, it will automatically install them.
}
\details{
The following packages will be installed if not already present:
\itemize{
\item "gdsfmt" (from Bioconductor repository)
\item "SeqArray" (from Bioconductor repository)
\item "SNPRelate" (from Bioconductor repository)
}
}
\examples{
\dontrun{data_processing_packages()}
}
