% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alazygtr.R
\name{do_everything_for_me}
\alias{do_everything_for_me}
\title{Lazy run of all landscape genomic analyses contained within \code{algatr}}
\usage{
do_everything_for_me(vcf, coords, envlayers)
}
\arguments{
\item{envlayers}{}
}
\description{
Disclaimer: this is probably a bad idea...
}
