% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{env_dist}
\alias{env_dist}
\title{Calculate environmental distance between environmental vars}
\usage{
env_dist(env)
}
\arguments{
\item{env}{dataframe with environmental values for each coordinate}
}
\description{
Calculate environmental distance between environmental vars
}
