% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{envirodata_packages}
\alias{envirodata_packages}
\title{Install environmental and geographic data processing packages}
\usage{
envirodata_packages()
}
\value{
None
}
\description{
Checks for the presence of packages required for environmental and geographic data processing.
If any of these packages are not already installed, it will automatically install them.
}
\details{
The following packages will be installed if not already present:
\itemize{
\item "RStoolbox" (from GitHub repository bleutner/RStoolbox)
\item "geodata"
\item "corrplot"
\item "vegan"
\item "gdistance"
\item "topoDistance"
\item "rmapshaper"
\item "wingen"(from GitHub repository AnushaPB/wingen)
}
}
\examples{
\dontrun{envirodata_packages()}
}
