% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_df}
\alias{gdm_df}
\title{Create dataframe of GDM results}
\usage{
gdm_df(gdm_result)
}
\arguments{
\item{gdm_result}{output of \link[algatr]{gdm_run}}
}
\value{
dataframe of gdm model coefficients
}
\description{
Create dataframe of GDM results
}
