% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_do_everything}
\alias{gdm_do_everything}
\title{GDM function to do everything (fit model, get coefficients, make and save raster)}
\usage{
gdm_do_everything(
  gendist,
  coords,
  envlayers = NULL,
  env = NULL,
  model = "best",
  sig = 0.05,
  nperm = 50,
  geodist_type = "Euclidean",
  dist_lyr = NULL,
  scale_gendist = FALSE,
  plot_vars = TRUE
)
}
\arguments{
\item{gendist}{matrix of genetic distances (must range between 0 and 1 or set scale_gendist = TRUE)}

\item{coords}{dataframe with x (i.e., longitude) and y (i.e., latitude) coordinates; must be in this order}

\item{envlayers}{envlayers for mapping (if env is provided, the dataframe column names and envlayers layer names should be the same)}

\item{env}{dataframe with environmental values for each coordinate; if not provided, it will be calculated based on coords/envlayers}

\item{model}{whether to fit the model with all variables ("full") or to perform variable selection to determine the best set of variables ("best"); (defaults to "best")}

\item{sig}{alpha value for significance threshold (defaults to 0.05); only used if model = "best"}

\item{nperm}{number of permutations to use to calculate variable importance; only used if model = "best" (defaults to 50)}

\item{geodist_type}{the type of geographic distance to be calculated; options are "Euclidean" (default) for direct distance, "topographic" for topographic distances, and "resistance" for resistance distances. Note: creation and plotting of the GDM raster is only possible for "Euclidean" distances}

\item{dist_lyr}{DEM raster for calculating topographic distances or resistance raster for calculating resistance distances}

\item{scale_gendist}{whether to scale genetic distance data from 0 to 1 (defaults to FALSE)}

\item{plot_vars}{whether to create variable vector loading plot (defaults to TRUE)}
}
\value{
list with final model, predictor coefficients, and PCA RGB map
}
\description{
GDM function to do everything (fit model, get coefficients, make and save raster)
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_var_select}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
