% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_doEverything}
\alias{gdm_doEverything}
\title{GDM function to do everything (fit model, get coefficients, make and save raster)
TODO: add default settings to below params}
\usage{
gdm_doEverything(
  gendist,
  coords,
  env = NULL,
  envlayers = NULL,
  model = "best",
  alpha = 0.05,
  nPerm = 50,
  scale = FALSE,
  plot_vars = TRUE
)
}
\arguments{
\item{gendist}{matrix of genetic distances (MUST RANGE BETWEEN 0 AND 1)}

\item{coords}{dataframe with x and y coordinates (MUST BE CALLED X AND Y) TODO:FIX THIS}

\item{env}{dataframe with environmental values for each coordinate, if not provided it will be calculated based on coords/envlayers}

\item{envlayers}{envlayers for mapping (MUST MATCH NAMES IN ENV DATAFRAME)}

\item{model}{whether to fit the model with all variables ("full") or to perform variable selection to determine the best set of variables ("best")}

\item{alpha}{alpha level for variable selection (defaults to 0.05), only matters if model = "best" TODO:ADD BETTER DESCRIPTOR}

\item{nPerm}{number of permutations to use to calculate variable importance, only matters if model = "best"}

\item{scale}{whether to scale genetic distance data from 0 to 1 TODO:ADD STOP}

\item{plot_vars}{whether to create variable vector loading plot}
}
\value{
list with final model, predictor coefficients, and PCA RGB map
}
\description{
GDM function to do everything (fit model, get coefficients, make and save raster)
TODO: add default settings to below params
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_var_select}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
