% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_format}
\alias{gdm_format}
\title{Format Data for Generalized Dissimilarity Modeling (GDM)}
\usage{
gdm_format(
  gendist,
  coords,
  env,
  scale_gendist = FALSE,
  geodist_type = "Euclidean",
  distPreds = NULL,
  dist_lyr = NULL,
  gdmPred = FALSE,
  gdmGen = FALSE
)
}
\arguments{
\item{gendist}{matrix of genetic distances (must range between 0 and 1 or set scale_gendist = TRUE)}

\item{coords}{dataframe with x (i.e., longitude) and y (i.e., latitude) coordinates; must be in this order}

\item{env}{dataframe or raster object with environmental values for each coordinate; if not provided, it will be calculated based on coords/envlayers}

\item{scale_gendist}{whether to scale genetic distance data from 0 to 1 (defaults to FALSE)}

\item{geodist_type}{the type of geographic distance to be calculated; options are "Euclidean" (default) for direct distance, "topographic" for topographic distances, and "resistance" for resistance distances. Note: creation and plotting of the GDM raster is only possible for "Euclidean" distances}

\item{dist_lyr}{DEM raster for calculating topographic distances or resistance raster for calculating resistance distances}

\item{gdmPred}{whether to include the gdm formatted predictor data seperately (defaults to FALSE). This dataframe contains the site number, coordinates, and environmental values at each site}

\item{gdmGen}{whether to include the gdm formatted genetic data seperately (defaults to FALSE). This dataframe contains the genetic distance matrix with an additional column for site number.}
}
\value{
either a gdmData object if gdmGen and gdmPred are FALSE or a list of gdm data objects
}
\description{
Format Data for Generalized Dissimilarity Modeling (GDM)
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_diss}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_table}()},
\code{\link{gdm_var_sel}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
