% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_map}
\alias{gdm_map}
\title{Make map from model}
\usage{
gdm_map(gdm_model, envlayers, coords, plot_vars = TRUE, scl = 1, plot = TRUE)
}
\arguments{
\item{gdm_model}{GDM model}

\item{envlayers}{stack of raster layers (NAMES MUST CORRESPOND WITH GDM MODEL)}

\item{plot_vars}{whether to create PCA plot to help in variable and map interpretation}

\item{scl}{constant for rescaling variable vectors for plotting}

\item{plot}{}
}
\value{
GDM RGB map
}
\description{
Make map from model
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_var_select}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
