% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{gdm_packages}
\alias{gdm_packages}
\title{Install GDM packages}
\usage{
gdm_packages()
}
\value{
None
}
\description{
Checks for the presence of packages required for GDM.
If any of these packages are not already installed, it will automatically install them.
}
\details{
The following packages will be installed if not already present:
\itemize{
\item "cowplot"
\item "gdm"
}
}
\examples{
\dontrun{gdm_packages()}
}
