% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_plot_diss}
\alias{gdm_plot_diss}
\title{Plot compositional dissimilarity spline plots}
\usage{
gdm_plot_diss(gdm_model)
}
\arguments{
\item{gdm_model}{GDM model}
}
\value{
two spline plots of compositional dissimilarity
}
\description{
generates two plots: a plot of the observed response data against raw ecological distance from the model, and a plot of the observed response against the predicted response from the model (after link function is applied)
}
\details{
code is modified from the \code{plot.gdm()} function in the gdm package (Fitzpatrick et al. 2022)
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_table}()},
\code{\link{gdm_var_sel}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
