% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_plot_vars}
\alias{gdm_plot_vars}
\title{Create a PCA plot for GDM}
\usage{
gdm_plot_vars(
  pcaSamp,
  pcaRast,
  pcaRastRGB,
  coords,
  x = "PC1",
  y = "PC2",
  scl = 1,
  display_axes = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{pcaSamp}{PCA results from running prcomp()}

\item{pcaRast}{raster PCA}

\item{pcaRastRGB}{raster PCA rescaled to RGB}

\item{coords}{dataframe with x and y coordinates}

\item{x}{x-axis PC}

\item{y}{y-axis PC}

\item{scl}{constant for rescaling variable vectors for plotting}

\item{display_axes}{whether to display axes}

\item{quiet}{whether to print output tables and figures (defaults to FALSE)}
}
\value{
GDM PCA plot
}
\description{
Create a PCA plot for GDM
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_diss}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_run}()},
\code{\link{gdm_table}()},
\code{\link{gdm_var_select}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
