% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_run}
\alias{gdm_run}
\title{Run GDM and return model object}
\usage{
gdm_run(
  gendist,
  coords,
  env,
  model = "best",
  sig = 0.05,
  nperm = 50,
  scale_gendist = FALSE,
  geodist_type = "Euclidean",
  distPreds = NULL,
  dist_lyr = NULL
)
}
\arguments{
\item{gendist}{matrix of genetic distances (must range between 0 and 1 or set scale_gendist = TRUE)}

\item{coords}{dataframe with x (i.e., longitude) and y (i.e., latitude) coordinates; must be in this order}

\item{env}{dataframe or raster object with environmental values for each coordinate; if not provided, it will be calculated based on coords/envlayers}

\item{model}{whether to fit the model with all variables ("full") or to perform variable selection to determine the best set of variables ("best"); (defaults to "best")}

\item{sig}{alpha value for significance threshold (defaults to 0.05); only used if model = "best"}

\item{nperm}{number of permutations to use to calculate variable importance; only used if model = "best" (defaults to 50)}

\item{scale_gendist}{whether to scale genetic distance data from 0 to 1 (defaults to FALSE)}

\item{geodist_type}{the type of geographic distance to be calculated; options are "Euclidean" (default) for direct distance, "topographic" for topographic distances, and "resistance" for resistance distances. Note: creation and plotting of the GDM raster is only possible for "Euclidean" distances}

\item{dist_lyr}{DEM raster for calculating topographic distances or resistance raster for calculating resistance distances}
}
\value{
GDM model
}
\description{
Run GDM and return model object
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_diss}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_table}()},
\code{\link{gdm_var_select}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
