% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_table}
\alias{gdm_table}
\title{Create \code{gt} table of GDM results}
\usage{
gdm_table(gdm_result, digits = 2, summary_stats = TRUE, footnote = TRUE)
}
\arguments{
\item{gdm_result}{output of \link[algatr]{gdm_run} or \link[algatr]{gdm_do_everything}}

\item{digits}{number of digits to include (defaults to 2)}
}
\value{
An object of class \code{gt_tbl}
}
\description{
Create \code{gt} table of GDM results
}
