% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_var_select}
\alias{gdm_var_select}
\title{Get best set of variables from a GDM model}
\usage{
gdm_var_select(gdmData, sig = 0.05, nperm = 10)
}
\arguments{
\item{gdmData}{data formatted using GDM package}

\item{sig}{sig level for determining variable significance}

\item{nperm}{number of permutations to run for variable testing}
}
\description{
Get best set of variables from a GDM model
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_diss}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_table}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
