% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_varimp}
\alias{gdm_varimp}
\title{Get best set of variables from a GDM model}
\usage{
gdm_varimp(gdmData, alpha = 0.05, nperm = 10, predSelect = TRUE)
}
\arguments{
\item{gdmData}{data formatted using GDM package}

\item{alpha}{alpha level for determining variable significance}

\item{nperm}{number of permutations to run for variable testing}
}
\description{
Get best set of variables from a GDM model
}
\seealso{
Other GDM functions: 
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
