% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_varimp_table}
\alias{gdm_varimp_table}
\title{Generate a Variable Importance Table for GDM Models}
\usage{
gdm_varimp_table(
  varimp,
  digits = 2,
  summary_stats = TRUE,
  nPerm = 50,
  geo = TRUE
)
}
\arguments{
\item{varimp}{a \code{gdmData} object or a variable importance object created by running \link[gdm]{gdm.varImp}.}

\item{digits}{number of digits to include (defaults to 2)}

\item{summary_stats}{whether to add summary statistics to bottom of table (defaults to TRUE).}

\item{nPerm}{number of permutations to use if \code{varimp} is a \code{gdmData} object. Default is 50.}

\item{geo}{whether to include geographic distance in the GDM model. Default is TRUE.}
}
\value{
A \code{gt} table object displaying the variable importance.
}
\description{
This function generates a table displaying the variable importance for Generalized Dissimilarity Models (GDM).
It can take either a \code{gdmData} object or a precomputed variable importance object and outputs a formatted table.
}
