% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dist.R
\name{gen_dist_calc}
\alias{gen_dist_calc}
\title{Calculate genetic distances}
\usage{
gen_dist_calc(
  vcf_file,
  plink_file,
  plink_id_file,
  dist_type,
  criticalpoint = 2.0234
)
}
\arguments{
\item{vcf_file}{path to vcf file}

\item{plink_file}{path to plink distance file (typically ".dist")}

\item{plink_id_file}{path to plink id file (typically ".dist.id")}

\item{dist_type}{is the type of genetic distance to calculate}

\item{criticalpoint}{is the critical point for the significance threshold for the TW test within the PCA}
}
\value{
pairwise distance matrix for given distance metric
}
\description{
Calculate genetic distances
}
