% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{gen_remove_na}
\alias{gen_remove_na}
\title{Remove NAs from genetic data
TODO: what is gen? gendist matrix? need to fill in param details}
\usage{
gen_remove_na(gen, nsd = 3)
}
\arguments{
\item{nsd}{}
}
\description{
Remove NAs from genetic data
TODO: what is gen? gendist matrix? need to fill in param details
}
