% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{gen_to_geno}
\alias{gen_to_geno}
\title{Convert dosage matrix or vcf to geno type object (N.B.: this only works for diploids!)}
\usage{
gen_to_geno(gen)
}
\arguments{
\item{gen}{a dosage matrix, an object of class 'vcfR', a path to a .vcf file, or an object of type snmfProject}
}
\value{
matrix encoded as geno type object
}
\description{
Convert dosage matrix or vcf to geno type object (N.B.: this only works for diploids!)
}
\seealso{
Other Imputation functions: 
\code{\link{geno_to_dosage}()},
\code{\link{simple_impute}()},
\code{\link{snmf_bestK}()},
\code{\link{snmf_crossent_helper}()},
\code{\link{str_impute}()}
}
\concept{Imputation functions}
