% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_dist.R
\name{genDist}
\alias{genDist}
\title{Function to measure genetic distances between samples, used as input for various downstream analyses
Genetic distances are measured by proportions of shared alleles (DPS)}
\usage{
genDist(genind, species)
}
\arguments{
\item{genind}{genind object containing genotypes (can contain missing values)}

\item{species}{specifies species name (Genus_species) to save output file}
}
\value{
square matrix with pairwise proportions of shared alleles (0 to 1), saved as csv
}
\description{
Function to measure genetic distances between samples, used as input for various downstream analyses
Genetic distances are measured by proportions of shared alleles (DPS)
}
