% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{genetic_distance_packages}
\alias{genetic_distance_packages}
\title{Install genetic distance packages}
\usage{
genetic_distance_packages()
}
\value{
None
}
\description{
Checks for the presence of packages required for genetic distance calculations.
If any of these packages are not already installed, it will automatically install them.
}
\details{
The following packages will be installed if not already present:
\itemize{
\item "adegenet"
\item "AssocTests"
\item "readr"
\item "tibble"
\item "ecodist"
\item "cowplot"
}
}
\examples{
\dontrun{genetic_distance_packages()}
}
