% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{geno_to_dosage}
\alias{geno_to_dosage}
\title{Convert lfmm/geno matrix to dosage matrix (N.B.: this only works for diploids!)}
\usage{
geno_to_dosage(geno)
}
\arguments{
\item{geno}{matrix of LEA geno or lfmm format (i.e., 0 corresponds to zero reference alleles)}
}
\value{
matrix encoded as dosage type object (0 corresponds to two reference alleles)
}
\description{
Convert lfmm/geno matrix to dosage matrix (N.B.: this only works for diploids!)
}
\seealso{
Other Imputation functions: 
\code{\link{gen_to_geno}()},
\code{\link{simple_impute}()},
\code{\link{snmf_bestK}()},
\code{\link{snmf_crossent_helper}()},
\code{\link{str_impute}()}
}
\concept{Imputation functions}
