% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{geo_dist}
\alias{geo_dist}
\title{Calculate geographic distance between sampling coordinates}
\usage{
geo_dist(coords)
}
\arguments{
\item{coords}{dataframe with x and y coordinates (MUST BE CALLED X AND Y)}
}
\description{
Calculate geographic distance between sampling coordinates
}
