% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_K.R
\name{get_K}
\alias{get_K}
\alias{get_K_tw}
\alias{get_K_elbow}
\alias{get_K_tess}
\title{Function to Select K value}
\usage{
get_K(gen, coords = NULL, k_selection = "tracy_widom", Kvals = Kvals, ...)

get_K_tw(gen)

get_K_elbow(gen)

get_K_tess(gen, coords, Kvals = 1:10, tess_method = "projected.ls", ploidy = 2)
}
\arguments{
\item{gen}{genotype matrix}

\item{coords}{coordinates for "tess"}

\item{k_selection}{method for performing K selection: "tracy_widom", "quick_elbow", or "tess"}

\item{Kvals}{values of K to test for "tess"}

\item{tess_method}{method to use for "tess"}

\item{ploidy}{ploidy for "tess"}
}
\description{
Function to Select K value
}
\section{Functions}{
\itemize{
\item \code{get_K_tw()}: Determine K using Tracy-Widom Test

\item \code{get_K_elbow()}: Determine K using PCA and \code{quick_elbow}

\item \code{get_K_tess()}: Determine K using TESS and \code{bestK}

}}
