% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{get_K_fc}
\alias{get_K_fc}
\title{Wrapper for find.clusters function}
\usage{
get_K_fc(gen, max.n.clust = nrow(x) - 1, perc.pca = 90, ...)
}
\arguments{
\item{gen}{a genotype matrix}

\item{max.n.clust}{ an \code{integer} indicating the maximum number of clusters
    to be tried. Values of 'k' will be picked up between 1 and
    \code{max.n.clust}}

\item{perc.pca}{a \code{numeric} value between 0 and 100 indicating the minimal
    percentage of the total variance of the data to be expressed by the retained
    axes of PCA.}

\item{...}{additional arguments to pass to \code{adegenet::find.clusters} function}
}
\description{
Wrapper for find.clusters function
}
