% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{get_loci}
\alias{get_loci}
\title{Get candidate loci row numbers}
\usage{
get_loci(pvec, sig)
}
\arguments{
\item{pvec}{vector of pvalues}

\item{sig}{alpha level}
}
\description{
Get candidate loci row numbers
}
