% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{ggbarplot_helper}
\alias{ggbarplot_helper}
\title{Helper function for TESS barplots using ggplot}
\usage{
ggbarplot_helper(dat)
}
\arguments{
\item{dat}{Q matrix}
}
\value{
barplot with Q-values and individuals, colorized by K-value
}
\description{
Helper function for TESS barplots using ggplot
}
