% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingen.R
\name{krig_agg_helper}
\alias{krig_agg_helper}
\title{Helper function for krig_helper; calculates aggregated/disaggregated raster cell size}
\usage{
krig_agg_helper(to_krig, agg_disagg, agg_spec = "agg")
}
\arguments{
\item{to_krig}{object to create grid for kriging, \code{grd} or \code{r} from \link[wingen]{krig_gd}}

\item{agg_disagg}{aggregation or disaggregation parameter, one of agg_grd, disagg_grd, agg_r, or disagg_r from \link[wingen]{krig_gd}}

\item{agg_spec}{whether aggregation or disaggregation is performed, options are \code{"disagg"} or \code{"agg"} (defaults to \code{"agg"})}
}
\value{
number of cells contained in final (aggregated or disaggregated) raster layer
}
\description{
Helper function for krig_helper; calculates aggregated/disaggregated raster cell size
}
\seealso{
Other wingen functions: 
\code{\link{krig_helper}()},
\code{\link{wingen_do_everything}()}
}
\concept{wingen functions}
