% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingen.R
\name{krig_helper}
\alias{krig_helper}
\title{Helper function for kriging wingen moving window; checks raster resolution and runs kriging}
\usage{
krig_helper(
  map,
  grd = NULL,
  index = 1,
  agg_grd = NULL,
  disagg_grd = NULL,
  agg_r = NULL,
  disagg_r = NULL
)
}
\arguments{
\item{map}{RasterLayer or RasterStack produced by \code{window_gd()}}

\item{grd}{object to create grid for kriging; inherits from \link[wingen]{krig_gd} (defaults to NULL)}

\item{index}{integer indices of layers in raster stack to krige; inherits from \link[wingen]{krig_gd} (defaults to 1)}

\item{agg_grd}{factor to use for aggregation of \code{grd}; inherits from \link[wingen]{krig_gd} (defaults to NULL)}

\item{disagg_grd}{factor to use for disaggregation of \code{grd}; inherits from \link[wingen]{krig_gd} (defaults to NULL)}

\item{agg_r}{factor to use for aggregation of \code{r}; inherits from \link[wingen]{krig_gd} (defaults to NULL)}

\item{disagg_r}{factor to use for disaggregation of \code{r}; inherits from \link[wingen]{krig_gd} (defaults to NULL)}
}
\value{
kriged map
}
\description{
Helper function for kriging wingen moving window; checks raster resolution and runs kriging
}
\seealso{
Other wingen functions: 
\code{\link{krig_agg_helper}()},
\code{\link{wingen_do_everything}()}
}
\concept{wingen functions}
