% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_LD.R
\name{ld_prune}
\alias{ld_prune}
\title{ld_prune prunes SNPs based on linkage disequilibrium using \code{snpggdsLDpruning()} in SNPRelate package.
TODO: if vcf is path convert, otherwise no need to convert (does functionality exist within snpgdsVCF2GDS?)
TODO: should intermediate gds files be deleted? set to temp and overwrite?
TODO: print number of SNPs to be retained after LD-pruning}
\usage{
ld_prune(
  vcf,
  outname,
  out_format,
  nodes = 1,
  ld.threshold = 0.6,
  slide.max.n = 100,
  maf = 0.05,
  seed = 1234,
  method = c("corr")
)
}
\arguments{
\item{vcf}{is the path to the vcf file containing all SNPs}

\item{outname}{prefix name of output files (will append with param settings)}

\item{out_format}{output file format ("plink" will produce ped and map files while "vcf" will produce a vcf and a GDS)
Saves SNP GDS and ped (plink) files with retained SNPs}

\item{nodes}{is the number of parallel processors (numeric)}

\item{ld.threshold}{is the threshold for LD pruning (numeric; 0 - 1; defaults to 0.6)}

\item{slide.max.n}{is the maximum number of SNPs in a sliding window (numeric; defaults to 100)}

\item{maf}{is the minor allele frequency cutoff (numeric; defaults to 0.05)}

\item{seed}{is the random starting seed (defaults to 1234)}

\item{method}{is the LD threshold method; default to corr which is r2 correlation coefficient}

\item{output_dir}{is the output file directory (default is here)}
}
\description{
ld_prune prunes SNPs based on linkage disequilibrium using \code{snpggdsLDpruning()} in SNPRelate package.
TODO: if vcf is path convert, otherwise no need to convert (does functionality exist within snpgdsVCF2GDS?)
TODO: should intermediate gds files be deleted? set to temp and overwrite?
TODO: print number of SNPs to be retained after LD-pruning
}
