% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_LD.R
\name{LDprune}
\alias{LDprune}
\title{LDprune prunes SNPs based on linkage disequilibrium using snpggdsLDpruning() in SNPRelate package.}
\usage{
LDprune(
  vcf,
  species,
  output_dir,
  nodes = 1,
  ld.threshold = 0.6,
  slide.max.n = 1000,
  maf = 0.05,
  seed = 1234,
  method = c("corr")
)
}
\arguments{
\item{vcf}{is the vcf file containing all SNPs}

\item{species}{is the species name; used for file naming}

\item{output_dir}{is the output file directory

Saves SNP GDS and ped (plink) files with retained SNPs as "Species_LDpruned_params"}

\item{nodes}{is the number of parallel processors (numeric)}

\item{ld.threshold}{is the threshold for LD pruning (numeric; 0 - 1)}

\item{slide.max.n}{is the maximum number of SNPs in a sliding window (numeric)}

\item{maf}{is the minor allele frequency cutoff (numeric)}

\item{seed}{is the random starting seed}

\item{method}{is the LD threshold method; corr is r2 correlation coefficient}
}
\description{
LDprune prunes SNPs based on linkage disequilibrium using snpggdsLDpruning() in SNPRelate package.
}
