% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_manhattanplot}
\alias{lfmm_manhattanplot}
\title{LFMM Manhattan Plot}
\usage{
lfmm_manhattanplot(pvalues, sig)
}
\arguments{
\item{pvalues}{pvalue dataframe}

\item{sig}{alpha level}
}
\description{
LFMM Manhattan Plot
}
