% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{lfmm_packages}
\alias{lfmm_packages}
\title{Install LFMM packages}
\usage{
lfmm_packages()
}
\value{
None
}
\description{
Checks for the presence of packages required for LFMM.
If any of these packages are not already installed, it will automatically install them.
}
\details{
The following packages will be installed if not already present:
\itemize{
\item "adegenet"
\item "AssocTests"
\item "lfmm"
\item "tess3r" (from GitHub repository bcm-uga/TESS3_encho_sen)
\item "LEA" (from Bioconductor repository)
}
}
\examples{
\dontrun{lfmm_packages()}
}
