% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_qqplot}
\alias{lfmm_qqplot}
\title{LFMM QQplot}
\usage{
lfmm_qqplot(df)
}
\arguments{
\item{df}{dataframe of LFMM test results produced by \code{lfmm_df}}
}
\value{
a QQ-plot
}
\description{
LFMM QQplot
}
\seealso{
Other LFMM functions: 
\code{\link{lfmm_df}()},
\code{\link{lfmm_do_everything}()},
\code{\link{lfmm_manhattanplot}()},
\code{\link{lfmm_run}()},
\code{\link{lfmm_table}()},
\code{\link{lfmm_test_tidy}()},
\code{\link{quick_elbow}()},
\code{\link{select_K}()}
}
\concept{LFMM functions}
