% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_table}
\alias{lfmm_table}
\title{Create \code{gt} table of LFMM results}
\usage{
lfmm_table(
  df,
  sig = 0.05,
  sig_only = TRUE,
  top = FALSE,
  order = FALSE,
  var = NULL,
  rows = NULL,
  digits = 2,
  footnotes = TRUE
)
}
\arguments{
\item{digits}{}
}
\description{
Create \code{gt} table of LFMM results
}
