% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_test_tidy}
\alias{lfmm_test_tidy}
\title{Helper function to tidy LFMM test results}
\usage{
lfmm_test_tidy(colname, lfmm_test_result)
}
\arguments{
\item{colname}{names of elements within lfmm_test_result}

\item{lfmm_test_result}{LFMM test results}
}
\description{
Helper function to tidy LFMM test results
}
\seealso{
Other LFMM functions: 
\code{\link{lfmm_df}()},
\code{\link{lfmm_do_everything}()},
\code{\link{lfmm_manhattanplot}()},
\code{\link{lfmm_qqplot}()},
\code{\link{lfmm_run}()},
\code{\link{lfmm_table}()},
\code{\link{quick_elbow}()},
\code{\link{select_K}()}
}
\concept{LFMM functions}
