% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liz_coords}
\alias{liz_coords}
\title{Example coordinates from Bouzid et al. 2022}
\format{
A dataframe with x and y coordinates for 53 samples
}
\source{
\url{https://doi.org/10.1111/mec.15836}
}
\usage{
liz_coords
}
\description{
Example coordinates from Bouzid et al. 2022
}
\keyword{datasets}
