% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liz_gendist}
\alias{liz_gendist}
\title{Example genetic distance matrix, calculated with Plink using data from Bouzid et al. 2022}
\format{
A dataframe with genetic distances for 53 samples
}
\source{
\url{https://doi.org/10.1111/mec.15836}
}
\usage{
liz_gendist
}
\description{
Example genetic distance matrix, calculated with Plink using data from Bouzid et al. 2022
}
\keyword{datasets}
