% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liz_vcf}
\alias{liz_vcf}
\title{Example VCF from Bouzid et al. 2022}
\format{
A vcfR object with 1000 loci and 53 samples
}
\source{
\url{https://doi.org/10.1111/mec.15836}
}
\usage{
liz_vcf
}
\description{
Example VCF from Bouzid et al. 2022
}
\keyword{datasets}
