% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{make_boundary_raster}
\alias{make_boundary_raster}
\title{Make raster of population boundaries}
\usage{
make_boundary_raster(
  krig_admix,
  K,
  m = matrix(rep(1, 9), nrow = 3),
  minQ = NULL,
  boundaries = FALSE
)
}
\arguments{
\item{krig_admix}{}

\item{K}{}

\item{m}{}

\item{minQ}{}

\item{boundaries}{}
}
\description{
Make raster of population boundaries
}
