% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{map_maxQ}
\alias{map_maxQ}
\title{Make raster with max Q values}
\usage{
map_maxQ(
  krig_admix,
  alpha = alpha,
  add = FALSE,
  cols = "Default",
  plot_map = FALSE
)
}
\arguments{
\item{krig_admix}{}

\item{alpha}{}

\item{add}{}

\item{cols}{}

\item{plot_map}{}

\item{zlim}{}
}
\description{
Make raster with max Q values
}
