% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{masking_packages}
\alias{masking_packages}
\title{Install masking packages}
\usage{
masking_packages()
}
\value{
None
}
\description{
Checks for the presence of packages required for masking.
If the package is not already installed, it will automatically install it.
}
\details{
The following package will be installed if not already present:
\itemize{
\item "geodata"
}
}
\examples{
\dontrun{masking_packages()}
}
