% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{MMRR.varSel}
\alias{MMRR.varSel}
\title{MMRR.varSel performs MMRR with backward elimination variable selection}
\usage{
MMRR.varSel(Y, X, nperm = 999, stdz = TRUE)
}
\arguments{
\item{Y}{is a dependent distance matrix}

\item{X}{is a list of independent distance matrices (with optional names)}

\item{nperm}{is the number of permutations to be used in significance tests. Default = 999.}

\item{stdz}{if TRUE then matrices will be standardized. Default = TRUE.}
}
\description{
MMRR.varSel performs MMRR with backward elimination variable selection
}
