% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{mmrr_df}
\alias{mmrr_df}
\title{Make nice dataframe from MMRR results}
\usage{
mmrr_df(mod)
}
\arguments{
\item{mod}{the fitted MMRR model}
}
\description{
Make nice dataframe from MMRR results
}
\seealso{
Other MMRR functions: 
\code{\link{MMRR}()},
\code{\link{mmrr_best}()},
\code{\link{mmrr_do_everything}()},
\code{\link{mmrr_full}()},
\code{\link{mmrr_plot}()},
\code{\link{mmrr_table}()},
\code{\link{mmrr_var_sel}()},
\code{\link{unfold}()}
}
\concept{MMRR functions}
