% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{mmrr_do_everything}
\alias{mmrr_do_everything}
\title{MMRR function to do everything ()
TODO: gendist matrix must range between 0 and 1 as with GDM?}
\usage{
mmrr_do_everything(
  gendist,
  coords,
  env = NULL,
  envlayers,
  model = "best",
  n_perm = 999
)
}
\arguments{
\item{gendist}{matrix of genetic distances}

\item{coords}{dataframe with x and y coordinates (MUST BE CALLED X AND Y)}

\item{env}{dataframe with environmental values for each coordinate, if not provided it will be calculated based on coords/envlayers}

\item{envlayers}{envlayers for mapping (MUST MATCH NAMES IN ENV DATAFRAME)}

\item{model}{whether to fit the model with all variables ("full") or to perform variable selection to determine the best set of variables ("best"); defaults to "best"}

\item{n_perm}{number of permutations to use to calculate variable importance, only matters if model = "best" (defaults to 999)}
}
\description{
MMRR function to do everything ()
TODO: gendist matrix must range between 0 and 1 as with GDM?
}
