% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{mmrr_packages}
\alias{mmrr_packages}
\title{Install MMRR packages}
\usage{
mmrr_packages()
}
\value{
None
}
\description{
Checks for the presence of packages required for MMRR.
If the package is not already installed, it will automatically install it.
}
\details{
The following package will be installed if not already present:
\itemize{
\item "GGally"
}
}
\examples{
\dontrun{mmrr_packages()}
}
