% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{mmrr_plot}
\alias{mmrr_plot}
\title{Plot MMRR results}
\usage{
mmrr_plot(
  Y = NULL,
  X,
  mod = NULL,
  plot_type = "all",
  stdz = TRUE,
  var_names = NULL
)
}
\arguments{
\item{Y}{The dependent variable in the form of a distance matrix}

\item{X}{A list of independent variables in the form of distance matrices (required if \code{plot_type = "fitted"}, \code{vars} or \code{"all"})}

\item{mod}{The fitted MMRR model (required if \code{plot_type = "fitted"} or \code{"all"})}

\item{plot_type}{which plots to produce (options: (1) "vars" to plot single variable relationships, (2) "fitted" to plot the fitted relationship, (3) "cov" to plot covariances between the predictor variables, (4) "all" to produce all plots (default))}

\item{scale}{If TRUE, all variables are scaled}

\item{varNames}{A vector of names for the variables in the model (optional)}
}
\description{
Plot MMRR results
}
