% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{mmrr_run}
\alias{mmrr_run}
\title{Run MMRR and return model object}
\usage{
mmrr_run(Y, X, nperm = 999, stdz = TRUE, model = "full")
}
\arguments{
\item{Y}{dependent distance matrix}

\item{X}{list of independent distance matrices (with optional names)}

\item{nperm}{number of permutations to use to calculate variable importance; only used if \code{model = "best"} (default = 999)}

\item{stdz}{if TRUE then matrices will be standardized (default = TRUE)}

\item{model}{whether to fit the model with all variables (\code{"full"}) or to perform variable selection to determine the best set of variables (\code{"best"}); defaults to "full"}
}
\value{
list with final model results and regression coefficients
}
\description{
Run MMRR and return model object
}
\seealso{
Other MMRR functions: 
\code{\link{MMRR}()},
\code{\link{mmrr_df}()},
\code{\link{mmrr_do_everything}()},
\code{\link{mmrr_plot}()},
\code{\link{mmrr_table}()},
\code{\link{mmrr_var_sel}()},
\code{\link{unfold}()}
}
\concept{MMRR functions}
