% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{mmrr_table}
\alias{mmrr_table}
\title{Create \code{gt} table of MMRR results}
\usage{
mmrr_table(mmrr_results, digits = 2, summary_stats = TRUE)
}
\arguments{
\item{digits}{}
}
\value{
An object of class \code{gt_tbl}
}
\description{
Create \code{gt} table of MMRR results
}
