% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{outliers}
\alias{outliers}
\title{Z outlier finder
from https://popgen.nescent.org/2018-03-27_RDA_GEA.html}
\usage{
outliers(x, z)
}
\description{
Z outlier finder
from https://popgen.nescent.org/2018-03-27_RDA_GEA.html
}
