% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{plot_boundaries}
\alias{plot_boundaries}
\title{Plot boundaries}
\usage{
plot_boundaries(
  rv_stack,
  option = "all",
  log_transform = FALSE,
  col = magma(100),
  weights = NULL,
  tess3_obj = NULL,
  main = ""
)
}
\arguments{
\item{rv_stack}{}

\item{option}{}

\item{log_transform}{}

\item{col}{}

\item{weights}{}

\item{tess3_obj}{}

\item{main}{}
}
\description{
Plot boundaries
}
