% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{plot_crossent}
\alias{plot_crossent}
\title{Helper function to plot cross entropy scores from SNMF}
\usage{
plot_crossent(ce_values)
}
\arguments{
\item{ce_values}{df with run, K-value, and cross entropy created in \link[algatr]{snmf_bestK}}
}
\value{
ggplots of cross entropy values compared to K-values (and across runs)
}
\description{
Helper function to plot cross entropy scores from SNMF
}
