% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot the relationship between two distance metrics
TODO: are metric names really necessary?}
\usage{
plot_dist(dist_x, dist_y, metric_name_x, metric_name_y)
}
\arguments{
\item{dist_x}{df containing square distance matrix for x axis}

\item{dist_y}{df containing square distance matrix for y axis}

\item{metric_name_x}{name of distance metric for x axis}

\item{metric_name_y}{name of distance metric for y axis}
}
\description{
Plot the relationship between two distance metrics
TODO: are metric names really necessary?
}
