% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{plot_extrap_mask}
\alias{plot_extrap_mask}
\title{Plot mask on top of map}
\usage{
plot_extrap_mask(
  map_r,
  map_mask,
  RGB_cols = TRUE,
  mask_col = rgb(0, 0, 0, alpha = 0.9)
)
}
\arguments{
\item{map_r}{SpatRaster or Raster* object to be masked}

\item{map_mask}{SpatRaster or Raster* object with 1s where you want to mask and NA everywhere else (i.e., what you want to keep, as produced by \code{\link{extrap_mask}})}

\item{RGB_cols}{whether the plot should be RGB-based or not}

\item{mask_col}{color and transparency of mask (defaults to black and alpha = 0.9)}
}
\value{
plot \code{map} with areas masked based on \code{map_mask}
}
\description{
Plots a raster with another raster "mask" on top of it
}
\seealso{
\code{\link{extrap_mask}}
}
