% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_rgb.R
\name{plot_rgb}
\alias{plot_rgb}
\title{RGB plot of a raster}
\usage{
plot_rgb(
  x,
  r = 1,
  g = 2,
  b = 3,
  a = NULL,
  scale = NULL,
  mar = 0,
  stretch = NULL,
  smooth = TRUE,
  colNA = "white",
  alpha = NULL,
  bgalpha = NULL,
  zlim = NULL,
  zcol = FALSE,
  axes = FALSE,
  ...
)
}
\arguments{
\item{x}{RasterStack or SpatRaster}

\item{r}{integer between 1 and \code{nlyr(x)}. Layer to use as the Red channel}

\item{g}{integer between 1 and \code{nlyr(x)}. Layer to use as the Green channel}

\item{b}{integer between 1 and \code{nlyr(x)}. Layer to use as the Blue channel}

\item{a}{NULL or integer between 1 and \code{nlyr(x)}. Layer to use as the alpha (transparency) channel. If not NULL, argument \code{alpha} is ignored}

\item{scale}{integer. Maximum (possible) value in the three channels. Defaults to 255 or to the maximum value of \code{x} if that is known and larger than 255}

\item{mar}{numeric vector recycled to length 4 to set the margins of the plot. Use \code{mar=NULL} or \code{mar=NA} to not set the margins}

\item{stretch}{character. Option to stretch the values to increase contrast: "lin" (linear) or "hist" (histogram)}

\item{smooth}{logical. If \code{TRUE}, smooth the image when drawing to get the appearance of a higher spatial resolution}

\item{colNA}{color. The color used for cells that have NA values}

\item{alpha}{transparency. Integer between 0 (transparent) and 255 (opaque)}

\item{bgalpha}{Background transparency. Integer between 0 (transparent) and 255 (opaque)}

\item{zlim}{numeric vector of length 2. Range of values to plot (optional). If this is set, and \code{stretch="lin"} is used, then the values are stretched within the range of \code{zlim}. This allows creating consistent coloring between SpatRasters with different cell-value ranges, even when stretching the colors for improved contrast}

\item{zcol}{logical. If \code{TRUE} the values outside the range of zlim get the color of the extremes of the range. Otherwise, the values outside the zlim range get the color of \code{NA} values (see argument "colNA")}

\item{axes}{logical. If \code{TRUE} axes are drawn (and arguments such as \code{main="title"} will be honored)}

\item{...}{graphical parameters as in \code{\link[terra]{plot}}<SpatRaster-method>}
}
\description{
Wrapper of the \link[terra]{plotRGB} function with added steps to perform
re-scaling of raster layers (e.g., from 0 to 255 for color layers).
}
