% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{plotMMRR}
\alias{plotMMRR}
\title{Plot MMMR}
\usage{
plotMMRR(reg, Y, X, scale = TRUE, varNames = NULL, lineCol = "blue", ...)
}
\arguments{
\item{reg}{The fitted MMRR model}

\item{Y}{The dependent variable in the form of a distance matrix}

\item{X}{A list of independent variables in the form of distance matrices (with optional names)}

\item{scale}{If TRUE, all variables are scaled}

\item{varNames}{A vector of names for the variables in the model (optional)}

\item{lineCol}{Color for regression line}

\item{...}{Additional arguments to be passed to plot() function (optional)}
}
\description{
Plots the results of an MMRR analysis
}
\details{
The objects supplied for Y and X should be the same variables used to fit the MMRR model.  The parameter 'scale' should be the same as used to fit the model.
The varNames argument can be used to specify variable names for labeling the plot axes.  The first name is for the dependent variable; additional names should be supplied in the same order as the independent variables.

When using MMRR, please cite the original citation:
Wang I.J. (2013) Examining the full effects of landscape heterogeneity on spatial genetic variation: a multiple matrix regression approach for quantifying geographic and ecological isolation. Evolution, 67: 3403-3411.
}
