% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{pops_helper}
\alias{pops_helper}
\title{Helper function to get population assignments for best K}
\usage{
pops_helper(gen, tess3_obj, K)
}
\arguments{
\item{gen}{genotype dosage matrix (rows = individuals & columns = SNPs) or \code{vcfR} object}

\item{tess3_obj}{list produced by \code{\link{tess3}}}

\item{K}{K value}
}
\value{
population assignments for each individual based on max Q values
}
\description{
Helper function to get population assignments for best K
}
\seealso{
Other TESS functions: 
\code{\link{allK_plot_helper}()},
\code{\link{bestK}()},
\code{\link{ggbarplot_helper}()},
\code{\link{krig_legend}()},
\code{\link{tess_barplot}()},
\code{\link{tess_col_default}()},
\code{\link{tess_do_everything}()},
\code{\link{tess_ggbarplot}()},
\code{\link{tess_ggplot}()},
\code{\link{tess_krig}()},
\code{\link{tess_ktest}()},
\code{\link{tess_plot_allK}()}
}
\concept{TESS functions}
