% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_pca.R
\name{raster_pca}
\alias{raster_pca}
\title{Perform Raster PCA}
\usage{
raster_pca(x, center = TRUE, scale = TRUE, tol = NULL, rank. = NULL, ...)
}
\arguments{
\item{x}{A Raster* or SpatRaster object}

\item{center}{a logical value indicating whether the variables
    should be shifted to be zero centered. Alternately, a vector of
    length equal the number of columns of \code{x} can be supplied.
    The value is passed to \code{scale}.}

\item{scale}{a logical value indicating whether the variables should be scaled to have unit variance before the analysis takes place. The default is TRUE, unlike for \link[stats]{prcomp}, since in most cases scaling should be performed.}

\item{tol}{a value indicating the magnitude below which components
    should be omitted. (Components are omitted if their
    standard deviations are less than or equal to \code{tol} times the
    standard deviation of the first component.)  With the default null
    setting, no components are omitted (unless \code{rank.} is specified
    less than \code{min(dim(x))}.).  Other settings for tol could be
    \code{tol = 0} or \code{tol = sqrt(.Machine$double.eps)}, which
    would omit essentially constant components.}

\item{rank.}{optionally, a number specifying the maximal rank, i.e.,
    maximal number of principal components to be used.  Can be set as
    alternative or in addition to \code{tol}, useful notably when the
    desired rank is considerably smaller than the dimensions of the matrix.}

\item{...}{arguments passed to or from other methods.  If \code{x} is
    a formula one might specify \code{scale.} or \code{tol}.}
}
\value{
A SpatRaster of principal component scores
}
\description{
Runs Principal Component Analysis (PCA) on a Raster stack.
}
\examples{
library(terra)

# Create a Raster stack with random data
set.seed(123)
r <- c(rast(matrix(rnorm(100), nrow = 10)),  rast(matrix(rnorm(100), nrow = 10)))

pca_result <- raster_pca(r)

}
