% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_biplot}
\alias{rda_biplot}
\title{Plot RDA biplot}
\usage{
rda_biplot(TAB_loci, TAB_var, rda_axes = c(1, 2))
}
\arguments{
\item{TAB_loci}{}

\item{TAB_var}{}

\item{rda_axes}{}
}
\description{
Plot RDA biplot
}
