% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_cor}
\alias{rda_cor}
\title{Test for associations}
\usage{
rda_cor(rda_gen, env)
}
\arguments{
\item{rda_gen}{genotype matrix for candidate snps}

\item{env}{env dataframe}
}
\description{
Test for associations
}
