% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_cor_env_helper}
\alias{rda_cor_env_helper}
\title{Helper function for rda_cor_test}
\usage{
rda_cor_env_helper(snp_name, snp_df, env)
}
\arguments{
\item{env}{}
}
\description{
Helper function for rda_cor_test
}
\keyword{internal}
