% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_cor_helper}
\alias{rda_cor_helper}
\title{Helper function for rda_cor_test}
\usage{
rda_cor_helper(envvar, snp)
}
\arguments{
\item{snp}{}
}
\description{
Helper function for rda_cor_test
}
\keyword{internal}
