% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_getsnps}
\alias{rda_getsnps}
\title{Run get SNPs}
\usage{
rda_getsnps(
  mod,
  naxes = "all",
  outlier_method = "p",
  padj_method = "fdr",
  sig = 0.05,
  z = 3,
  plot = TRUE
)
}
\description{
Run get SNPs
}
