% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_lm_env_helper}
\alias{rda_lm_env_helper}
\title{Helper function for rda_lm_test}
\usage{
rda_lm_env_helper(locus, env)
}
\arguments{
\item{locus}{}

\item{env}{}
}
\description{
Helper function for rda_lm_test
}
\keyword{internal}
