% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_lm_helper}
\alias{rda_lm_helper}
\title{Helper function for rda_lm_test}
\usage{
rda_lm_helper(envvar, locus)
}
\arguments{
\item{envvar}{}

\item{locus}{}
}
\description{
Helper function for rda_lm_test
}
\keyword{internal}
