% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_lm_test}
\alias{rda_lm_test}
\title{Test for associations}
\usage{
rda_lm_test(rda_gen, env)
}
\arguments{
\item{rda_gen}{genotype matrix for candidate loci}

\item{env}{env dataframe}
}
\description{
Test for associations
}
