% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_loadings}
\alias{rda_loadings}
\title{Generates plots of loadings from RDA model}
\usage{
rda_loadings(mod, naxes = NULL)
}
\arguments{
\item{mod}{model object of class \code{rda}}

\item{naxes}{number of axes to display loadings (defaults to the number of enviro. vars in call; values that exceed this will display loadings along PC axes)}
}
\value{
frequency histograms with loading of each SNP along each RDA axis
}
\description{
Generates plots of loadings from RDA model
}
\seealso{
Other RDA functions: 
\code{\link{rda_cor}()},
\code{\link{rda_do_everything}()},
\code{\link{rda_getoutliers}()},
\code{\link{rda_plot}()},
\code{\link{rda_run}()},
\code{\link{rda_table}()},
\code{\link{rda_varpart_table}()},
\code{\link{rda_varpart}()}
}
\concept{RDA functions}
