% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_manhattan}
\alias{rda_manhattan}
\title{RDA manhattan plot}
\usage{
rda_manhattan(TAB_loci, pvalues, sig)
}
\arguments{
\item{TAB_loci}{}

\item{pvalues}{}

\item{sig}{}
}
\description{
RDA manhattan plot
}
