% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{rda_packages}
\alias{rda_packages}
\title{Install RDA packages}
\usage{
rda_packages()
}
\value{
None
}
\description{
Checks for the presence of packages required for RDA.
If any of these packages are not already installed, it will automatically install them.
}
\details{
The following packages will be installed if not already present:
\itemize{
\item "ggrepel"
\item "qvalue" (from Bioconductor repository)
\item "robust"
\item "tibble"
\item "vegan"
}
}
\examples{
\dontrun{rda_packages()}
}
