% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_plot}
\alias{rda_plot}
\title{Plotting function}
\usage{
rda_plot(
  mod,
  rda_snps,
  pvalues = NULL,
  axes = "all",
  biplot_axes = NULL,
  sig = 0.05,
  manhattan = TRUE,
  rdaplot = TRUE,
  binwidth = NULL
)
}
\description{
Plotting function
}
