% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_run}
\alias{rda_run}
\title{Run RDA}
\usage{
rda_run(
  gen,
  env,
  coords = NULL,
  model = "full",
  correctGEO = FALSE,
  correctPC = FALSE,
  nPC = 3,
  Pin = 0.05,
  R2permutations = 1000,
  R2scope = T
)
}
\value{
RDA model
}
\description{
Run RDA
}
\seealso{
Other RDA functions: 
\code{\link{rda_cor}()},
\code{\link{rda_do_everything}()},
\code{\link{rda_getoutliers}()},
\code{\link{rda_loadings}()},
\code{\link{rda_plot}()},
\code{\link{rda_table}()},
\code{\link{rda_varpart_table}()},
\code{\link{rda_varpart}()}
}
\concept{RDA functions}
