% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_table}
\alias{rda_table}
\title{Make pretty table of p-values}
\usage{
rda_table(
  cor_df,
  sig = 0.05,
  sig_only = TRUE,
  top = FALSE,
  order = FALSE,
  var = NULL,
  nrow = NULL,
  digits = 2
)
}
\arguments{
\item{lm_df}{}
}
\description{
Make pretty table of p-values
}
