% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_tidy}
\alias{rda_tidy}
\title{Make dataframe for ggplot from RDA results}
\usage{
rda_tidy(mod, rda_snps, axes)
}
\arguments{
\item{rda_snps}{rda candidate snps}

\item{axes}{axes to include}

\item{rda_mod}{rda model}
}
\description{
Make dataframe for ggplot from RDA results
}
