% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_varpart_helper}
\alias{rda_varpart_helper}
\title{Helper function for \code{rda_varpart()}}
\usage{
rda_varpart_helper(mod)
}
\arguments{
\item{mod}{RDA model results}
}
\value{
df with relevant statistics (call, R2, adjusted R2, inertia)
}
\description{
Extracts relevant statistics from variance partitioning analysis
}
