% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_varpart_table}
\alias{rda_varpart_table}
\title{Create \code{gt} table with RDA variance partitioning results}
\usage{
rda_varpart_table(df, results, digits = 2, call_col = FALSE)
}
\arguments{
\item{df}{dataframe of variance partitioning results output from \link[algatr]{rda_varpart}}

\item{results}{results from \link[algatr]{rda_varpart}}

\item{digits}{number of digits to include (defaults to 2)}

\item{call_col}{whether to include column with RDA call (defaults to FALSE)}
}
\description{
Create \code{gt} table with RDA variance partitioning results
}
