% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{snmf_bestK}
\alias{snmf_bestK}
\title{Helper function to select "best" K based on minimizing cross-entropy criteria from SNMF results}
\usage{
snmf_bestK(snmf_proj, K, quiet)
}
\arguments{
\item{snmf_proj}{object of type snmfProject}

\item{K}{integer corresponding to K-value}

\item{quiet}{whether to print results of cross-entropy scores (defaults to TRUE; only does so if more than one K-value); only displays run with minimum cross-entropy}
}
\value{
list with best K-value and run number
}
\description{
Helper function to select "best" K based on minimizing cross-entropy criteria from SNMF results
}
\seealso{
Other Imputation functions: 
\code{\link{gen_to_geno}()},
\code{\link{geno_to_dosage}()},
\code{\link{simple_impute}()},
\code{\link{snmf_crossent_helper}()},
\code{\link{str_impute}()}
}
\concept{Imputation functions}
