% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{snmf_crossent_helper}
\alias{snmf_crossent_helper}
\title{Helper function to retrieve cross entropy scores from SNMF project}
\usage{
snmf_crossent_helper(snmf_proj, K, select_min = TRUE)
}
\arguments{
\item{snmf_proj}{object of type snmfProject}

\item{K}{K-value(s)}

\item{select_min}{whether to return minimum}
}
\value{
cross entropy scores for given K
}
\description{
Helper function to retrieve cross entropy scores from SNMF project
}
\seealso{
Other Imputation functions: 
\code{\link{gen_to_geno}()},
\code{\link{geno_to_dosage}()},
\code{\link{simple_impute}()},
\code{\link{snmf_bestK}()},
\code{\link{str_impute}()}
}
\concept{Imputation functions}
\keyword{internal}
