% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{str_impute}
\alias{str_impute}
\title{Imputation of missing values using population structure inferred with \code{LEA::snmf}}
\usage{
str_impute(
  gen,
  K,
  entropy = TRUE,
  repetitions = 10,
  project = "new",
  quiet = TRUE,
  save_output = FALSE,
  output_filename = NULL
)
}
\arguments{
\item{gen}{a dosage matrix, an object of class 'vcfR', a path to a .vcf file, or an object of type snmfProject}

\item{K}{
    An integer vector corresponding to the number of ancestral populations for
    which the snmf algorithm estimates have to be calculated.
}

\item{entropy}{
    A boolean value. If true, the cross-entropy criterion is calculated 
    (see \code{\link[LEA]{create.dataset}} and 
    \code{\link[LEA]{cross.entropy.estimation}}).
}

\item{repetitions}{
    An integer corresponding with the number of repetitions for each value of 
    \code{K}.
}

\item{project}{
    A character string among "continue", "new", and "force". If "continue",
    the results are stored in the current project. If "new", the current 
    project is removed and a new one is created to store the result. If 
    "force", the results are stored in the current project even if the input 
    file has been modified since the creation of the project.
}

\item{quiet}{whether to print results of cross-entropy scores (defaults to TRUE; only does so if more than one K-value); only displays run with minimum cross-entropy}

\item{save_output}{if TRUE, saves SNP GDS and ped (plink) files with retained SNPs in new directory; if FALSE returns object (defaults to TRUE)}

\item{output_filename}{if \code{save_output = TRUE}, name prefix for saved .geno file, SNMF project file, and SNMF output file results (defaults to FALSE, in which no files are saved)}
}
\value{
dosage matrix with imputed missing values
}
\description{
Imputation of missing values using population structure inferred with \code{LEA::snmf}
}
\seealso{
Other Imputation functions: 
\code{\link{gen_to_geno}()},
\code{\link{geno_to_dosage}()},
\code{\link{simple_impute}()},
\code{\link{snmf_bestK}()},
\code{\link{snmf_crossent_helper}()}
}
\concept{Imputation functions}
