% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_barplot}
\alias{tess_barplot}
\title{Create TESS barplot}
\usage{
tess_barplot(
  qmat,
  col_pal = algatr_col_default("base"),
  sort_by_Q = TRUE,
  legend = TRUE,
  legend_position = "bottomright",
  border = NA,
  space = 0,
  ...
)
}
\arguments{
\item{qmat}{Q matrix}

\item{sort_by_Q}{whether to sort bars by Q value (equivalent to \link[tess3r]{barplot} sort.by.Q)}

\item{legend}{whether to display legend (defaults to TRUE)}

\item{legend_position}{the x and y coordinates or keyword to determine legend position (defaults to bottom right)}

\item{border}{the color to be used for the border of the bars.
    Use \code{border = NA} to omit borders.  If there are shading
    lines, \code{border = TRUE} means use the same colour for
    the border as for the shading lines.}

\item{space}{the amount of space (as a fraction of the average bar
    width) left before each bar.  May be given as a single number or
    one number per bar.  If \code{height} is a matrix and
    \code{beside} is \code{TRUE}, \code{space} may be specified by
    two numbers, where the first is the space between bars in the
    same group, and the second the space between the groups.  If not
    given explicitly, it defaults to \code{c(0,1)} if \code{height}
    is a matrix and \code{beside} is \code{TRUE}, and to 0.2
    otherwise.}

\item{...}{other parameters of the function \code{\link{barplot.default}}.}
}
\value{
STRUCTURE-style bar plot of TESS results
}
\description{
Based on code from: https://github.com/bcm-uga/TESS3_encho_sen/blob/master/R/plotQ.R
}
\seealso{
Other TESS functions: 
\code{\link{allK_plot_helper}()},
\code{\link{bestK}()},
\code{\link{krig_legend}()},
\code{\link{tess_col_default}()},
\code{\link{tess_do_everything}()},
\code{\link{tess_ggplot}()},
\code{\link{tess_krig}()},
\code{\link{tess_ktest}()},
\code{\link{tess_plot_allK}()}
}
\concept{TESS functions}
