% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_col_default}
\alias{tess_col_default}
\title{Create default TESS color palette}
\usage{
tess_col_default(n, alpha = 1)
}
\arguments{
\item{n}{number of colors to generate (must be less than 9)}
}
\description{
Create default TESS color palette
}
\seealso{
Other TESS functions: 
\code{\link{allK_plot_helper}()},
\code{\link{bestK}()},
\code{\link{tess_barplot}()},
\code{\link{tess_do_everything}()},
\code{\link{tess_ggplot}()},
\code{\link{tess_krig}()},
\code{\link{tess_ktest}()},
\code{\link{tess_plot_allK}()}
}
\concept{TESS functions}
