% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_full}
\alias{tess_full}
\title{TESS function to do everything}
\usage{
tess_full(
  gen,
  coords,
  spdf,
  Kvals = 1:10,
  ploidy = 2,
  n_cell = 1000,
  k_selection = "auto",
  plot_method = "max",
  minQ = 0,
  tess_method = "projected.ls"
)
}
\arguments{
\item{gen}{genotype matrix}

\item{coords}{dataframe with x and y coordinates}

\item{spdf}{Spatial Points Dataframe to create grid for kriging}

\item{Kvals}{vector of K values to test}

\item{ploidy}{ploidy of organism (defaults to 2)}

\item{n_cell}{number of grid cells to use when kriging}

\item{k_selection}{how to perform k selection (options: "auto" for automatic selection based on \code{\link{bestK}} or "manual" to enter into console)}

\item{plot_method}{method for making rainbow map of kriged layers (options: "max" to only plot the max Q value for each cell or "all" to plot all Qvalues greater than \code{minQ})}

\item{minQ}{threshold for minimum Q-value for rainbow plotting if \code{method = "all"} is used}
}
\value{
list with all TESS results, final K value, and final kriged raster
}
\description{
TESS function to do everything
}
\concept{TESS functions}
