% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_ggbarplot}
\alias{tess_ggbarplot}
\title{Create TESS barplot using ggplot2}
\usage{
tess_ggbarplot(
  qmat,
  ggplot_fill = algatr_col_default("ggplot"),
  sort_by_Q = TRUE,
  legend = TRUE
)
}
\arguments{
\item{qmat}{Q matrix}

\item{ggplot_fill}{any ggplot2 scale fill discrete function (default: \link[algatr]{scale_fill_viridis_d}, \code{option = "turbo"})}

\item{sort_by_Q}{whether to sort bars by Q value (equivalent to \link[tess3r]{barplot} sort.by.Q)}

\item{legend}{whether to display legend (defaults to TRUE)}
}
\value{
ggplot object of TESS results as a barplot
}
\description{
Create TESS barplot using ggplot2
}
