% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_ktest}
\alias{tess_ktest}
\title{Test multiple K values}
\usage{
tess_ktest(
  gen,
  coords,
  Kvals = 1:10,
  grid = NULL,
  tess_method = "projected.ls",
  lambda = 1,
  K_selection = "manual",
  ploidy = 2,
  quiet = FALSE
)
}
\arguments{
\item{gen}{genotype dosage matrix (rows = individuals & columns = snps) or \code{vcfR} object}

\item{coords}{coordinates of samples as sf points, a two-column matrix, or a data.frame representing x and y coordinates (see Details for important information about projections)}

\item{Kvals}{vector of K values to test}

\item{grid}{Optional SpatRaster for kriging (defaults to NULL)}

\item{tess_method}{the type of TESS method to be run ("projected.ls" for projected least squares algorithm (default) or "qp" for quadratic programming algorithm)}

\item{lambda}{numeric value for the spatial regularization parameter. The default value lambda = 1 attributes equal weights to the loss function and to the penalty function.}

\item{K_selection}{how to perform K selection ("manual" to enter into console (default) or "auto" for automatic selection based on \link[algatr]{bestK})}

\item{ploidy}{ploidy of data (defaults to 2)}

\item{quiet}{whether to operate quietly and suppress the output of tables and figures (defaults to FALSE)}
}
\value{
list with results from testing different K-values
}
\description{
Test multiple K values
}
\seealso{
Other TESS functions: 
\code{\link{allK_plot_helper}()},
\code{\link{bestK}()},
\code{\link{geom_tess}()},
\code{\link{ggbarplot_helper}()},
\code{\link{pops_helper}()},
\code{\link{tess_barplot}()},
\code{\link{tess_col_default}()},
\code{\link{tess_do_everything}()},
\code{\link{tess_ggbarplot}()},
\code{\link{tess_ggplot}()},
\code{\link{tess_krig}()},
\code{\link{tess_legend}()},
\code{\link{tess_plot_allK}()}
}
\concept{TESS functions}
