% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_ktest_plot}
\alias{tess_ktest_plot}
\title{Plot results from K value testing}
\usage{
tess_ktest_plot(tess_results, n_cell = 1000, plot_method = "max")
}
\arguments{
\item{tess_results}{list produced by \code{\link{tess_ktest}} with Kvals, coords, and spdf}

\item{n_cell}{number of grid cells to use when kriging}

\item{plot_method}{method for making rainbow map of kriged layers (options: "max" to only plot the max Q value for each cell or "all" to plot all Qvalues greater than \code{minQ})}
}
\description{
Plot results from K value testing
}
