% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{tess_packages}
\alias{tess_packages}
\title{Install TESS packages}
\usage{
tess_packages()
}
\value{
None
}
\description{
Checks for the presence of packages required for TESS.
If any of these packages are not already installed, it will automatically install them.
}
\details{
The following packages will be installed if not already present:
\itemize{
\item "automap"
\item "graphics"
\item "LEA" (from Bioconductor repository)
\item "tess3r" (from GitHub repository bcm-uga/TESS3_encho_sen)
\item "fields"
\item "rworldmap"
\item "cowplot"
}
}
\examples{
\dontrun{tess_packages()}
}
