% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_plot}
\alias{tess_plot}
\title{Make rainbow TESS plot from kriged admixture plots}
\usage{
tess_plot(
  krig_admix,
  coords = NULL,
  plot_method = "maxQ",
  col_pal = viridis::turbo,
  col_breaks = 20,
  col_alpha = 0.5,
  minQ = 0.1
)
}
\arguments{
\item{krig_admix}{RasterStack returned by \link[landgen]{tess_krig}}

\item{coords}{dataframe with x and y coordinates for plotting (optional)}

\item{plot_method}{method for making rainbow map of kriged layers (options: "maxQ" to only plot the max Q value for each cell (default), "allQ" to plot all Qvalues greater than \code{minQ}, "maxQ_poly" or "allQ_poly" to create the plots as previously described, but as polygons for each K instead of continuous Q values)}

\item{col_pal}{function that creates a vector of contiguous colors (defaults to \link[viridis]{turbo})}

\item{col_breaks}{if using "maxQ" and "allQ" plot methods, the number of breaks to use when plotting kriged maps}

\item{col_alpha}{if using the "allQ" plot method, an alpha-transparency level in the range \link{0,1} (0 means transparent and 1 means opaque) provided to \code{col_pal} function}

\item{minQ}{threshold for minimum Q-value for rainbow plotting if \code{method = "all"} is used (defaults to 0.10)}
}
\description{
Make rainbow TESS plot from kriged admixture plots
}
