% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_rainbow_plot}
\alias{tess_rainbow_plot}
\alias{tess_plot_max}
\alias{tess_plot_all}
\title{Make rainbow TESS plot from kriged admixture plots}
\usage{
tess_rainbow_plot(
  krig_admix,
  coords = NULL,
  spdf = NULL,
  plot_method = "max",
  minQ = 0.2,
  alpha = 1,
  col_pal = "Default",
  reclassify = FALSE,
  poly = FALSE
)

tess_plot_max(krig_admix, K = K, reclassify = FALSE, poly = FALSE, col_pal)

tess_plot_all(
  krig_admix,
  K = K,
  reclassify = FALSE,
  poly = FALSE,
  minQ = 0.1,
  col_pal
)
}
\arguments{
\item{krig_admix}{}

\item{coords}{dataframe with x and y coordinates for plotting (not required)}

\item{spdf}{Spatial Points Dataframe for plotting (not required)}

\item{plot_method}{method for making rainbow map of kriged layers (options: "max" to only plot the max Q value for each cell or "all" to plot all Qvalues greater than \code{minQ})}

\item{minQ}{}

\item{alpha}{transparency for plotting}

\item{col_pal}{}

\item{reclassify}{}

\item{poly}{}

\item{K}{}
}
\description{
Make rainbow TESS plot from kriged admixture plots

Plot method: Q max

Plot method: Q all
}
\section{Functions}{
\itemize{
\item \code{tess_plot_max()}: plot based on max of Qvalues

\item \code{tess_plot_all()}: plot based on all Qvalues

}}
