% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_reclass}
\alias{tess_reclass}
\title{Reclassify Kriged Qmatrix Raster}
\usage{
tess_reclass(r, inc = 0.05, rec = "Default")
}
\arguments{
\item{r}{kriged Q raster}

\item{inc}{increment for reclassification bins}

\item{rec}{reclassify matrix (if not provided, will be set to default)}
}
\description{
Reclassify Kriged Qmatrix Raster
}
