% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{unfold}
\alias{unfold}
\title{unfold converts the lower diagonal elements of a matrix into a vector}
\usage{
unfold(X, scale = TRUE)
}
\arguments{
\item{X}{is a distance matrix.}

\item{scale}{if TRUE then matrices will be standardized. Default = TRUE.}
}
\description{
unfold converts the lower diagonal elements of a matrix into a vector
}
