% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installs.R
\name{wingen_packages}
\alias{wingen_packages}
\title{Install wingen packages}
\usage{
wingen_packages()
}
\value{
None
}
\description{
This function checks for the presence of packages required for wingen.
If the packages are not already installed, it will automatically install them.
}
\details{
The following package will be installed if not already present:
\itemize{
\item "wingen"
}
}
\examples{
\dontrun{wingen_packages()}
}
