% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R, R/TESS.R
\name{bestK}
\alias{bestK}
\title{Best K Selection based on cross entropy}
\usage{
bestK(tess3_obj, Kvals)

bestK(tess3_obj, Kvals)
}
\arguments{
\item{tess3_obj}{list produced by \code{\link{tess3}}}

\item{Kvals}{vector of K values for testing}
}
\description{
Best K Selection based on cross entropy

Best K Selection based on cross entropy
}
\note{
(source: https://chazhyseni.github.io/NALgen/post/determining_bestk/)

(source: https://chazhyseni.github.io/NALgen/post/determining_bestk/)
}
