% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vars.R
\name{check_dists}
\alias{check_dists}
\title{Check geographic and environmental distances for collinearity}
\usage{
check_dists(envlayers, coords, type = "Euclidean", lyr = NULL)
}
\arguments{
\item{envlayers}{A RasterStack of data layers}

\item{coords}{Dataframe with x and y sample coordinates.}

\item{type}{The type of geographic distance to be calculated; options are "Euclidean" for direct distance, "topographic" for topographic distances, and "resistance" for resistance distances.}

\item{lyr}{DEM raster for calculating topographic distances or resistance raster for calculating resistance distances}
}
\value{
A list with matrices of p-values and Mantel's r
}
\description{
Check geographic and environmental distances for collinearity
}
