% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vars.R
\name{check_env}
\alias{check_env}
\title{Check environmental layers for collinearity}
\usage{
check_env(envlayers, threshold = 0.7)
}
\arguments{
\item{envlayers}{SpatRaster or Raster* object}

\item{threshold}{the cutoff correlation coefficient for flagging variables as collinear (numeric; defaults to 0.7)}
}
\value{
a matrix of correlation coefficients
}
\description{
Check environmental layers for collinearity
}
