% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vars.R
\name{check_vals}
\alias{check_vals}
\title{Check extracted values for collinearity}
\usage{
check_vals(envlayers, coords, threshold = 0.7)
}
\arguments{
\item{envlayers}{A RasterStack of data layers}

\item{coords}{Dataframe with x and y sample coordinates.}

\item{threshold}{The cutoff correlation coefficient for flagging variables as collinear (numeric)}
}
\value{
A matrix of correlation coefficients
}
\description{
Check extracted values for collinearity
}
