% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alazygtr.R
\name{do_everything_for_me}
\alias{do_everything_for_me}
\title{Lazy run of all landscape genomic analyses contained within \code{algatr}}
\usage{
do_everything_for_me(gen, coords, envlayers, quiet = FALSE, gators = FALSE)
}
\arguments{
\item{gen}{path to vcf file, a \code{vcfR} type object, or a dosage matrix}

\item{coords}{dataframe with x (i.e., longitude) and y (i.e., latitude) coordinates; must be in this order}

\item{envlayers}{SpatRaster or Raster* for mapping (if env is provided, the dataframe column names and envlayers layer names should be the same)}

\item{quiet}{whether to print output tables and figures (defaults to FALSE)}

\item{gators}{set to TRUE to see some gators...}
}
\value{
results from all six analyses contained within algatr
}
\description{
Disclaimer: this is probably a bad idea...
}
