% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{env_dist}
\alias{env_dist}
\title{Calculate distance between environmental vars}
\usage{
env_dist(env, stdz = TRUE)
}
\arguments{
\item{env}{dataframe or vector of environmental variables for locations}

\item{stdz}{if TRUE then environmental values will be standardized (default = TRUE)}
}
\value{
list of environmental distances between samples (for each environmental variable)
}
\description{
Calculate distance between environmental vars
}
