% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_coeffs}
\alias{gdm_coeffs}
\title{Get coefficients for each predictor}
\usage{
gdm_coeffs(gdm_model)
}
\arguments{
\item{gdm_model}{model of type \code{gdm}}
}
\value{
dataframe of coefficients for GDM
}
\description{
Get coefficients for each predictor
}
\seealso{
Other GDM functions: 
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_format}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_diss}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_table}()},
\code{\link{gdm_var_sel}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
