% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_df}
\alias{gdm_df}
\title{Get coefficients for each predictor}
\usage{
gdm_df(gdm_model)
}
\arguments{
\item{gdm_model}{\itemize{
\item model of type gdm
}}
}
\value{
data frame of coefficients for GDM
}
\description{
Get coefficients for each predictor
}
\seealso{
Other GDM functions: 
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_varimp}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
