% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_df}
\alias{gdm_df}
\title{Create dataframe of GDM results}
\usage{
gdm_df(gdm_result)
}
\arguments{
\item{gdm_result}{output of \link[algatr]{gdm_run}}
}
\value{
dataframe of gdm model coefficients
}
\description{
Create dataframe of GDM results
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_diss}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_table}()},
\code{\link{gdm_var_select}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
