% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_map}
\alias{gdm_map}
\title{Make map from model}
\usage{
gdm_map(
  gdm_model,
  envlayers,
  coords,
  plot_vars = TRUE,
  scl = 1,
  display_axes = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{gdm_model}{GDM model}

\item{envlayers}{SpatRaster or Raster* object (LAYER NAMES MUST CORRESPOND WITH GDM MODEL)}

\item{coords}{data frame with x and y coordinates}

\item{plot_vars}{whether to create PCA plot to help in variable and map interpretation (defaults to TRUE)}

\item{scl}{constant for rescaling variable vectors for plotting (defaults to 1)}

\item{display_axes}{display PC axes text, labels, and ticks (defaults to FALSE)}

\item{quiet}{whether to operate quietly and suppress the output of tables and figures (defaults to FALSE)}
}
\value{
GDM RGB map
}
\description{
Make map from model
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_format}()},
\code{\link{gdm_plot_diss}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_table}()},
\code{\link{gdm_var_sel}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
