% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_plot_isplines}
\alias{gdm_plot_isplines}
\title{Plot isplines for each variable}
\usage{
gdm_plot_isplines(gdm_model)
}
\arguments{
\item{gdm_model}{GDM model}
}
\value{
plot for each I-spline
}
\description{
Plot isplines for each variable
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_var_select}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
