% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_run}
\alias{gdm_run}
\title{Run GDM and return model object}
\usage{
gdm_run(
  gendist,
  coords,
  env,
  model = "best",
  alpha = 0.05,
  nperm = 50,
  scale = FALSE,
  geodist_type = "Euclidean",
  distPreds = NULL,
  dist_lyr = NULL
)
}
\arguments{
\item{gendist}{matrix of genetic distances (MUST RANGE BETWEEN 0 AND 1)}

\item{coords}{dataframe with x and y coordinates}

\item{env}{dataframe with environmental values for each coordinate}

\item{model}{whether to compute the full model ("full") or the best model based on variable selection steps ("best")}

\item{alpha}{alpha level for variable selection (defaults to 0.05), only matters if model = "best"}

\item{nperm}{number of permutations to use to calculate variable importance, only matters if model = "best"}

\item{scale}{scale genetic distance data from 0 to 1}
}
\value{
GDM model
}
\description{
Run GDM and return model object
}
\seealso{
Other GDM functions: 
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_varimp}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
