% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_table}
\alias{gdm_table}
\title{Make publication quality table of GDM results}
\usage{
gdm_table(gdm_df, digits = 2)
}
\arguments{
\item{gdm_df}{dataframe output from \link[algatr]{gdm_df}}

\item{digits}{number of digits to include (defaults to 2)}
}
\value{
An object of class \code{gt_tbl}
}
\description{
Make publication quality table of GDM results
}
