% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{gdm_table}
\alias{gdm_table}
\title{Create \code{gt} table of GDM results}
\usage{
gdm_table(gdm_result, digits = 2, summary_stats = TRUE, footnote = TRUE)
}
\arguments{
\item{gdm_result}{output of \link[algatr]{gdm_run} or \link[algatr]{gdm_do_everything}}

\item{digits}{number of digits to include (defaults to 2)}
}
\value{
An object of class \code{gt_tbl}
}
\description{
Create \code{gt} table of GDM results
}
\seealso{
Other GDM functions: 
\code{\link{gdm_coeffs}()},
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_diss}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_var_select}()},
\code{\link{scale01}()}
}
\concept{GDM functions}
