% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_dist.R
\name{gen_dist}
\alias{gen_dist}
\title{Calculate genetic distances}
\usage{
gen_dist(
  vcf = NULL,
  plink_file = NULL,
  plink_id_file = NULL,
  dist_type,
  criticalpoint = 2.0234
)
}
\arguments{
\item{vcf}{path to vcf file or a \code{vcfR} type object}

\item{plink_file}{path to plink distance file (typically ".dist")}

\item{plink_id_file}{path to plink id file (typically ".dist.id")}

\item{dist_type}{the type of genetic distance to calculate}

\item{criticalpoint}{the critical point for the significance threshold for the TW test within the PCA}
}
\value{
pairwise distance matrix for given distance metric
}
\description{
Calculate genetic distances
}
\details{
Euclidean and Bray-Curtis distances calculated using the ecodist package: Goslee, S.C. and Urban, D.L. 2007. The ecodist package for dissimilarity-based analysis of ecological data. Journal of Statistical Software 22(7):1-19. DOI:10.18637/jss.v022.i07.
Proportions of shared alleles calculated using the adegenet package: Jombart T. and Ahmed I. (2011) adegenet 1.3-1: new tools for the analysis of genome-wide SNP data. Bioinformatics. doi:10.1093/bioinformatics/btr521.
For PC-based distances, mean-based imputation calculated using the dartR package: Gruber, B, Unmack, PJ, Berry, OF, Georges, A. dartr: An r package to facilitate analysis of SNP data generated from reduced representation genome sequencing. Mol Ecol Resour. 2018; 18: 691-699. https://doi.org/10.1111/1755-0998.12745
}
