% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dist.R
\name{gen_dist_hm}
\alias{gen_dist_hm}
\title{Make heatmap of genetic distances}
\usage{
gen_dist_hm(dist)
}
\arguments{
\item{dist}{Matrix of genetic distances}
}
\value{
heatmap of genetic distances
}
\description{
Make heatmap of genetic distances
}
