% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{ggbarplot_helper}
\alias{ggbarplot_helper}
\title{Helper function for TESS barplots using ggplot}
\usage{
ggbarplot_helper(dat)
}
\arguments{
\item{dat}{Q matrix}
}
\value{
barplot with Q-values and individuals, colorized by K-value
}
\description{
Helper function for TESS barplots using ggplot
}
\seealso{
Other TESS functions: 
\code{\link{allK_plot_helper}()},
\code{\link{bestK}()},
\code{\link{krig_legend}()},
\code{\link{pops_helper}()},
\code{\link{tess_barplot}()},
\code{\link{tess_col_default}()},
\code{\link{tess_do_everything}()},
\code{\link{tess_ggbarplot}()},
\code{\link{tess_ggplot}()},
\code{\link{tess_krig}()},
\code{\link{tess_ktest}()},
\code{\link{tess_plot_allK}()}
}
\concept{TESS functions}
