% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_do_everything}
\alias{lfmm_do_everything}
\title{LFMM function to do everything}
\usage{
lfmm_do_everything(
  gen,
  env,
  coords = NULL,
  K = NULL,
  lfmm_method = "ridge",
  K_selection = "tracy_widom",
  Kvals = 1:10,
  sig = 0.05,
  p_adj = "fdr",
  calibrate = "gif",
  criticalpoint = 2.0234,
  low = 0.08,
  max.pc = 0.9,
  perc.pca = 90,
  max.n.clust = 10,
  quiet = FALSE
)
}
\arguments{
\item{gen}{genotype dosage matrix (rows = individuals & columns = snps) or \code{vcfR} object}

\item{env}{dataframe with environmental data or a Raster* type object from which environmental values for the coordinates can be extracted}

\item{coords}{dataframe with coordinates (only needed if K selection is performed with TESS or if environmental values aren't provided)}

\item{K}{number of latent factors (if left as NULL (default), K value selection will be conducted)}

\item{lfmm_method}{lfmm method (either \code{"ridge"} (default) or \code{"lasso"})}

\item{K_selection}{method for performing k selection (can either by "tracy_widom" (default), "quick_elbow", "tess", or "find_clusters")}

\item{Kvals}{values of K to test for "tess"}

\item{sig}{alpha level for determining candidate SNPs (defaults to 0.05)}

\item{p_adj}{method to use for p-value correction (defaults to "fdr"); other options can be found in \code{\link{p.adjust}}}

\item{calibrate}{a character string, "gif" or "median+MAD". If the "gif" option is set (default),
significance values are calibrated by using the genomic control method. Genomic control
uses a robust estimate of the variance of z-scores called "genomic inflation factor".
If the "median+MAD" option is set, the pvalues are calibrated by computing the median and MAD of the zscores. If \code{NULL}, the
pvalues are not calibrated.}

\item{criticalpoint}{if "tracy_widom" method is used, a numeric value corresponding to the significance level. If the significance level is 0.05, 0.01, 0.005, or 0.001, the criticalpoint should be set to be 0.9793, 2.0234, 2.4224, or 3.2724, respectively (defaults to 2.0234)}

\item{low}{if "quick_elbow" method is used, numeric, between zero and one, the threshold that defines whether a principal component explains 'much' of the variance (defaults to 0.08).}

\item{max.pc}{if "quick_elbow" method is used, maximum percentage of the variance to capture before the elbow (cumulative sum to PC 'n'; defaults to 0.90).}

\item{perc.pca}{if "find_clusters" method is used, a numeric value between 0 and 100 indicating the minimal percentage of the total variance of the data to be expressed by the retained axes of PCA (defaults to 90).}

\item{max.n.clust}{if "find_clusters" method is used, an integer indicating the maximum number of clusters to be tried. Values of 'k' will be picked up between 1 and max.n.clust (defaults to 10)}

\item{quiet}{whether to print output tables and figures (defaults to FALSE)}
}
\value{
list with candidate SNPs, model results, and K-value
}
\description{
LFMM function to do everything
}
\details{
LFMM is run using the lfmm package: Jumentier, B. (2021). lfmm: Latent Factor Mixed Models. R package version 1.1.
See also: Caye, K., Jumentier, B., Lepeule, J., & François, O. (2019). LFMM 2: Fast and accurate inference of gene-environment associations in genome-wide studies. Mol. Biol. Evol. 36(4):852-860. doi: https://doi.org/10.1093/molbev/msz008
}
\seealso{
Other LFMM functions: 
\code{\link{lfmm_df}()},
\code{\link{lfmm_manhattanplot}()},
\code{\link{lfmm_qqplot}()},
\code{\link{lfmm_run}()},
\code{\link{lfmm_table}()},
\code{\link{lfmm_test_tidy}()},
\code{\link{quick_elbow}()},
\code{\link{select_K}()}
}
\concept{LFMM functions}
