% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_manhattanplot}
\alias{lfmm_manhattanplot}
\title{LFMM Manhattan Plot}
\usage{
lfmm_manhattanplot(df, sig, group = NULL, var = NULL)
}
\arguments{
\item{df}{dataframe of lfmm test results produced by \code{lfmm_df}}

\item{sig}{significance cutoff}
}
\value{
Manhattan plot
}
\description{
LFMM Manhattan Plot
}
\seealso{
Other LFMM functions: 
\code{\link{lfmm_df}()},
\code{\link{lfmm_do_everything}()},
\code{\link{lfmm_qqplot}()},
\code{\link{lfmm_run}()},
\code{\link{lfmm_table}()},
\code{\link{lfmm_test_tidy}()},
\code{\link{quick_elbow}()},
\code{\link{select_K}()}
}
\concept{LFMM functions}
