% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_manhattanplot}
\alias{lfmm_manhattanplot}
\title{LFMM Manhattan Plot}
\usage{
lfmm_manhattanplot(df, sig, group = NULL, var = NULL)
}
\arguments{
\item{df}{dataframe of lfmm test results produced by \link[wingen]{lfmm_df}}

\item{sig}{significance cutoff}
}
\description{
LFMM Manhattan Plot
}
