% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_qqplot}
\alias{lfmm_qqplot}
\title{LFMM QQplot}
\usage{
lfmm_qqplot(df)
}
\arguments{
\item{df}{dataframe of LFMM test results produced by \code{lfmm_df}}
}
\value{
a QQ-plot
}
\description{
LFMM QQplot
}
