% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_run}
\alias{lfmm_run}
\title{Run LFMM}
\usage{
lfmm_run(
  gen,
  env,
  coords = NULL,
  K = NULL,
  lfmm_method = "ridge",
  k_selection = "tracy.widom",
  Kvals = 1:20,
  sig = 0.05
)
}
\arguments{
\item{gen}{genotype matrix}

\item{env}{dataframe with environmental data}

\item{coords}{dataframe with coordinates (only needed if K selection is performed with TESS)}

\item{K}{number of latent factors (if left blank, K value selection will be conducted)}

\item{lfmm_method}{lfmm method (either \code{"ridge"} or \code{"lasso"})}

\item{k_selection}{method for performing k selection (can either by "tracy.widom", "quick.elbow", or "tess")}

\item{Kvals}{values of K to test if using "tess" method of K selection}

\item{sig}{alpha level for determining candidate loci}
}
\description{
Run LFMM
}
