% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_run}
\alias{lfmm_run}
\title{Run LFMM}
\usage{
lfmm_run(
  gen,
  env,
  K,
  lfmm_method = "ridge",
  p_adj = "fdr",
  sig = 0.05,
  calibrate = "gif"
)
}
\arguments{
\item{gen}{genotype dosage matrix (rows = individuals & columns = snps) or \code{vcfR} object}

\item{env}{dataframe with environmental data or a Raster* type object from which environmental values for the coordinates can be extracted}

\item{K}{number of latent factors (if left as NULL (default), K value selection will be conducted)}

\item{lfmm_method}{lfmm method (either \code{"ridge"} (default) or \code{"lasso"})}

\item{p_adj}{method to use for p-value correction (defaults to "none")}

\item{sig}{alpha level for determining candidate SNPs (defaults to 0.05)}

\item{calibrate}{a character string, "gif" or "median+MAD". If the "gif" option is set (default),
significance values are calibrated by using the genomic control method. Genomic control
uses a robust estimate of the variance of z-scores called "genomic inflation factor".
If the "median+MAD" option is set, the pvalues are calibrated by computing the median and MAD of the zscores. If \code{NULL}, the
pvalues are not calibrated.}
}
\description{
Run LFMM
}
