% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMM.R
\name{lfmm_table}
\alias{lfmm_table}
\title{Title}
\usage{
lfmm_table(
  df,
  sig = 0.05,
  sig_only = TRUE,
  top = FALSE,
  order = FALSE,
  var = NULL,
  nrow = NULL,
  digits = 2
)
}
\arguments{
\item{digits}{}
}
\description{
Title
}
