% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{MMRR}
\alias{MMRR}
\title{MMRR performs Multiple Matrix Regression with Randomization analysis}
\usage{
MMRR(Y, X, nperm = 999, stdz = TRUE)
}
\arguments{
\item{Y}{is a dependent distance matrix}

\item{X}{is a list of independent distance matrices (with optional names)}

\item{nperm}{is the number of permutations to be used in significance tests. Default = 999.}

\item{stdz}{if TRUE then matrices will be standardized. Default = TRUE.}
}
\description{
MMRR performs Multiple Matrix Regression with Randomization analysis
}
