% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{mmrr_do_everything}
\alias{mmrr_do_everything}
\title{MMRR function to do everything}
\usage{
mmrr_do_everything(
  gendist,
  coords,
  env,
  geo = TRUE,
  model = "full",
  geodist_type = "Euclidean",
  dist_lyr = NULL,
  nperm = 999,
  stdz = TRUE,
  quiet = FALSE,
  plot_type = "all"
)
}
\arguments{
\item{gendist}{matrix of genetic distances}

\item{coords}{dataframe with x and y coordinates}

\item{env}{dataframe with environmental data or a Raster* type object from which environmental values for the coordinates can be extracted}

\item{geo}{whether to include geographic, topographic, or resistance distance as an independent variable (defaults to TRUE)}

\item{model}{whether to fit the model with all variables (\code{"full"}) or to perform variable selection to determine the best set of variables (\code{"best"}); defaults to "full"}

\item{geodist_type}{if \code{geo = TRUE}, the type of geographic distance to be calculated; options are "Euclidean" (default) for direct distance, "topographic" for topographic distances, and "resistance" for resistance distances}

\item{dist_lyr}{if \code{geodist_type = "topographic"}, DEM raster for calculating topographic distances or if \code{geodist_type = "resistance"}, resistance raster for calculating resistance distances}

\item{nperm}{number of permutations to be used in significance tests (default = 999)}

\item{stdz}{if TRUE then matrices will be standardized (default = TRUE)}

\item{quiet}{whether to plot results (default = FALSE)}

\item{plot_type}{if \code{quiet = FALSE}, which plots to produce (options: (1) "vars" to plot single variable relationships, (2) "fitted" to plot the fitted relationship, (3) "cov" to plot covariances between the predictor variables, (4) "all" to produce all plots (default))}
}
\value{
list with final model results and regression coefficients
}
\description{
MMRR function to do everything
}
\details{
The MMRR method is described here: Wang, I.J. (2013). Examining the full effects of landscape heterogeneity on spatial genetic variation: a multiple matrix regression approach for quantifying geographic and ecological isolation. Evolution 67(12):3403-3411. doi: https://doi.org/10.1111/evo.12134
}
\seealso{
Other MMRR functions: 
\code{\link{MMRR}()},
\code{\link{mmrr_df}()},
\code{\link{mmrr_plot}()},
\code{\link{mmrr_run}()},
\code{\link{mmrr_table}()},
\code{\link{mmrr_var_sel}()},
\code{\link{unfold}()}
}
\concept{MMRR functions}
