% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{mmrr_plot}
\alias{mmrr_plot}
\title{Plot MMRR results}
\usage{
mmrr_plot(
  Y = NULL,
  X,
  mod = NULL,
  plot_type = "all",
  stdz = TRUE,
  var_names = NULL
)
}
\arguments{
\item{Y}{the dependent variable in the form of a distance matrix}

\item{X}{a list of independent variables in the form of distance matrices (required if \code{plot_type = "fitted"}, \code{vars} or \code{"all"})}

\item{mod}{the fitted MMRR model (required if \code{plot_type = "fitted"} or \code{"all"})}

\item{plot_type}{which plots to produce (options: (1) "vars" to plot single variable relationships, (2) "fitted" to plot the fitted relationship, (3) "cov" to plot covariances between the predictor variables, (4) "all" to produce all plots (default))}

\item{stdz}{if TRUE then matrices will be standardized (default = TRUE)}

\item{var_names}{add variable names to plot (defaults to NULL)}
}
\value{
plots of MMRR single variable relationships, the fitted relationship, and the covariances between predictor variables
}
\description{
Plot MMRR results
}
\seealso{
Other MMRR functions: 
\code{\link{MMRR}()},
\code{\link{mmrr_df}()},
\code{\link{mmrr_do_everything}()},
\code{\link{mmrr_run}()},
\code{\link{mmrr_table}()},
\code{\link{mmrr_var_sel}()},
\code{\link{unfold}()}
}
\concept{MMRR functions}
