% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{mmrr_table}
\alias{mmrr_table}
\title{Create \code{gt} table of MMRR results}
\usage{
mmrr_table(mmrr_results, digits = 2, summary_stats = TRUE)
}
\arguments{
\item{mmrr_results}{results from MMRR}

\item{digits}{the number of decimal places to round to}

\item{summary_stats}{whether to add summary statistics (R-squared, F-statistic, F p-value) to bottom of table (defaults to TRUE)}
}
\value{
an object of class \code{gt_tbl}
}
\description{
Create \code{gt} table of MMRR results
}
\seealso{
Other MMRR functions: 
\code{\link{MMRR}()},
\code{\link{mmrr_df}()},
\code{\link{mmrr_do_everything}()},
\code{\link{mmrr_plot}()},
\code{\link{mmrr_run}()},
\code{\link{mmrr_var_sel}()},
\code{\link{unfold}()}
}
\concept{MMRR functions}
