% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_cor}
\alias{rda_cor}
\title{Genotype-environment correlation test}
\usage{
rda_cor(gen, var)
}
\arguments{
\item{gen}{dosage matrix}

\item{var}{dataframe with predictor variables}
}
\value{
dataframe with r and p-values from correlation test
}
\description{
Genotype-environment correlation test
}
\seealso{
Other RDA functions: 
\code{\link{rda_do_everything}()},
\code{\link{rda_getoutliers}()},
\code{\link{rda_plot}()},
\code{\link{rda_run}()},
\code{\link{rda_table}()},
\code{\link{rda_varpart}()},
\code{\link{rda_varpart_table}()}
}
\concept{RDA functions}
