% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_cor}
\alias{rda_cor}
\title{Genotype-environement correlation test}
\usage{
rda_cor(gen, var)
}
\arguments{
\item{gen}{dosage matrix}

\item{var}{dataframe with predictor variables}
}
\value{
dataframe with r and p-values from correlation test
}
\description{
Genotype-environement correlation test
}
