% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_do_everything}
\alias{rda_do_everything}
\title{RDA function to do everything}
\usage{
rda_do_everything(
  gen,
  env,
  coords = NULL,
  impute = "structure",
  K_impute = 3,
  entropy = TRUE,
  repetitions = 10,
  project = "new",
  quiet_impute = TRUE,
  save_output = FALSE,
  output_filename = NULL,
  model = "full",
  correctGEO = FALSE,
  correctPC = FALSE,
  outlier_method = "p",
  sig = 0.05,
  z = 3,
  p_adj = "fdr",
  cortest = TRUE,
  nPC = 3,
  varpart = FALSE,
  naxes = "all",
  Pin = 0.05,
  R2permutations = 1000,
  R2scope = T,
  stdz = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{gen}{genotype dosage matrix (rows = individuals & columns = SNPs) or \code{vcfR} object}

\item{env}{dataframe with environmental data or a Raster* type object from which environmental values for the coordinates can be extracted}

\item{coords}{dataframe with coordinates (only needed if correctGEO = TRUE) or if env is a Raster* from which values should be extracted}

\item{impute}{if NAs in \code{gen}, imputation will be performed on missing values; options are "structure" which uses the \code{str_impute()} function to impute based on population structure inferred with \code{LEA::snmf} (default); other option is "simple" based on \code{simple_impute()} which imputes to the median}

\item{K_impute}{if \code{impute = "structure"}, an integer vector (range or single value) corresponding to the number of ancestral populations for which the sNMF algorithm estimates have to be calculated (defaults to 3)}

\item{entropy}{
    A boolean value. If true, the cross-entropy criterion is calculated 
    (see \code{\link[LEA]{create.dataset}} and 
    \code{\link[LEA]{cross.entropy.estimation}}).
}

\item{repetitions}{
    An integer corresponding with the number of repetitions for each value of 
    \code{K}.
}

\item{project}{
    A character string among "continue", "new", and "force". If "continue",
    the results are stored in the current project. If "new", the current 
    project is removed and a new one is created to store the result. If 
    "force", the results are stored in the current project even if the input 
    file has been modified since the creation of the project.
}

\item{quiet_impute}{if \code{impute = "structure"}, whether to suppress results of cross-entropy scores (defaults to TRUE; only does so if K is range of values); only displays run with minimum cross-entropy}

\item{save_output}{if \code{impute = "structure"}, if TRUE, saves SNP GDS and ped (plink) files with retained SNPs in new directory; if FALSE returns object (defaults to FALSE)}

\item{output_filename}{if \code{impute = "structure"} and \code{save_output = TRUE}, name prefix for saved .geno file, SNMF project file, and SNMF output file results (defaults to FALSE, in which no files are saved)}

\item{model}{whether to fit the model with all variables ("full") or to perform variable selection to determine the best set of variables ("best"); defaults to "full"}

\item{correctGEO}{whether to condition on geographic coordinates}

\item{correctPC}{whether to condition on PCs from PCA of genotypes}

\item{outlier_method}{method to determine outliers. Can either be "p" to use the p-value method from \href{https://github.com/Capblancq/RDA-landscape-genomics}{here} or "z" to use the z-score based method from \href{https://popgen.nescent.org/2018-03-27_RDA_GEA.html}{here}}

\item{sig}{if \code{outlier_method = "p"}, the significance level to use to identify SNPs (defaults to 0.05)}

\item{z}{if \code{outlier_method = "z"}, the number of standard deviations to use to identify SNPs (defaults to 3)}

\item{p_adj}{if \code{outlier_method = "p"}, method to use for p-value correction (defaults to "fdr"); other options can be found in \code{\link[=p.adjust]{p.adjust()}}}

\item{cortest}{whether to create table of correlations for SNPs and environmental variable (defaults to TRUE)}

\item{nPC}{number of PCs to use if correctPC = TRUE (defaults to 3); if set to "manual" a selection option with a terminal prompt will be provided}

\item{varpart}{whether to perform variance partitioning (defaults to FALSE)}

\item{naxes}{number of RDA axes to use (defaults to "all" to use all axes), if set to "manual" a selection option with a terminal prompt will be given, otherwise can be any integer that is less than or equal to the total number of axes}

\item{Pin}{if \code{model = "best"}, limits of permutation P-values for adding (\code{Pin}) a term to the model, or dropping (\code{Pout}) from the model. Term is added if\code{ P <= Pin}, and removed if \code{P > Pout} (see \link[vegan]{ordiR2step}) (defaults to 0.05)}

\item{R2permutations}{if \code{model = "best"}, number of permutations used in the estimation of adjusted R2 for cca using RsquareAdj (see \link[vegan]{ordiR2step}) (defaults to 1000)}

\item{R2scope}{if \code{model = "best"} and set to TRUE (default), use adjusted R2 as the stopping criterion: only models with lower adjusted R2 than scope are accepted (see \link[vegan]{ordiR2step})}

\item{stdz}{whether to center and scale environmental data (defaults to TRUE)}

\item{quiet}{whether to operate quietly and suppress the output of tables and figures (defaults to FALSE)}
}
\value{
list containing (1) outlier SNPs, (2) dataframe with correlation test results, if \code{cortest = TRUE}, (3) the RDA model, (4) results from outlier analysis (output from \link[algatr]{rda_getoutliers}), (5) RDA R-Squared, (6) RDA ANOVA, (7) p-values if \code{outlier_method = "p"}, and (8) results from variance partitioning analysis, if \code{varpart = TRUE}
}
\description{
RDA function to do everything
}
\details{
Much of algatr's code is adapted from Capblancq T., Forester B.R. 2021. Redundancy analysis: A swiss army knife for landscape genomics. Methods Ecol. Evol. 12:2298-2309. doi: https://doi.org/10.1111/2041-210X.13722.
}
\seealso{
Other RDA functions: 
\code{\link{rda_cor}()},
\code{\link{rda_getoutliers}()},
\code{\link{rda_plot}()},
\code{\link{rda_run}()},
\code{\link{rda_table}()},
\code{\link{rda_varpart_table}()},
\code{\link{rda_varpart}()}
}
\concept{RDA functions}
