% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_do_everything}
\alias{rda_do_everything}
\title{RDA function to do everything}
\usage{
rda_do_everything(
  gen,
  env,
  coords = NULL,
  model = "best",
  correctGEO = FALSE,
  correctPC = FALSE,
  outlier_method = "p",
  sig = 0.05,
  z = 3,
  p_adj = "fdr",
  cortest = TRUE,
  nPC = 3,
  naxes = "all",
  Pin = 0.05,
  R2permutations = 1000,
  R2scope = T,
  stdz = TRUE
)
}
\arguments{
\item{gen}{genotype dosage matrix (rows = individuals & columns = SNPs) or \code{vcfR} object}

\item{env}{dataframe with environmental data or a Raster* type object from which environmental values for the coordinates can be extracted}

\item{coords}{dataframe with coordinates (only needed if correctGEO = TRUE) or if env is a Raster* from which values should be extracted}

\item{model}{whether to fit the model with all variables ("full") or to perform variable selection to determine the best set of variables ("best"); defaults to "best"}

\item{correctGEO}{whether to condition on geographic coordinates}

\item{correctPC}{whether to condition on PCs from PCA of genotypes}

\item{outlier_method}{method to determine outliers. Can either be "p" to use the p-value method from \href{https://github.com/Capblancq/RDA-landscape-genomics}{here} or "z" to use the z-score based method from \href{https://popgen.nescent.org/2018-03-27_RDA_GEA.html}{here}}

\item{sig}{if \code{outlier_method = "p"}, the significance level to use to identify SNPs (defaults to 0.05)}

\item{z}{if \code{outlier_method = "z"}, the number of standard deviations to use to identify SNPs (defaults to 3)}

\item{p_adj}{if \code{outlier_method = "p"}, method to use for p-value correction (defaults to "fdr"); other options can be found in \code{\link[=p.adjust]{p.adjust()}}}

\item{cortest}{whether to create table of correlations for SNPs and environmental variable (defaults to TRUE)}

\item{nPC}{number of PCs to use if correctPC = TRUE (defaults to 3); if set to "manual" a selection option with a terminal prompt will be provided}

\item{naxes}{number of RDA axes to use (defaults to "all" to use all axes), if set to "manual" a selection option with a terminal prompt will be given, otherwise can be any integer that is less than or equal to the total number of axes}

\item{Pin}{if \code{model = "best"}, limits of permutation P-values for adding (\code{Pin}) a term to the model, or dropping (\code{Pout}) from the model. Term is added if\code{ P <= Pin}, and removed if \code{P > Pout} (see \link[vegan]{ordiR2step})}

\item{R2permutations}{if \code{model = "best"}, number of permutations used in the estimation of adjusted R2 for cca using RsquareAdj (see \link[vegan]{ordiR2step})}

\item{R2scope}{if \code{model = "best"}, use adjusted R2 as the stopping criterion: only models with lower adjusted R2 than scope are accepted (see \link[vegan]{ordiR2step})}

\item{stdz}{whether to center and scale environmental data (defaults to TRUE)}
}
\value{
list containing (1) outlier SNPs, (2) dataframe with correlation test results, if \code{cortest = TRUE}, (3) the RDA model, (4) results from outlier analysis (output from \link[algatr]{rda_getoutliers}), (5) RDA R-Squared, (6) RDA ANOVA, (7) p-values if \code{outlier_method = "p"},
}
\description{
RDA function to do everything
}
