% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_getoutliers}
\alias{rda_getoutliers}
\title{Get significant outliers from RDA model}
\usage{
rda_getoutliers(
  mod,
  naxes = "all",
  outlier_method = "p",
  padj_method = "fdr",
  sig = 0.05,
  z = 3,
  plot = TRUE
)
}
\arguments{
\item{naxes}{number of RDA axes to use (defaults to "all" to use all axes), if set to "manual" a selection option with a terminal prompt will be given, otherwise can be any integer that is less than or equal to the total number of axes}

\item{outlier_method}{method to determine outliers. Can either be "p" to use the p-value method from https://github.com/Capblancq/RDA-landscape-genomics or "z" to use the z-score based method from https://popgen.nescent.org/2018-03-27_RDA_GEA.html}

\item{padj_method}{if \code{outlier_method = "p"}, the correction method supplied to \code{p.adjust} (can be "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{sig}{if \code{outlier_method = "p"}, the significance level to use to identify snps}

\item{z}{if \code{outlier_method = "z"}, the number of standard deviations to use to identify snps}

\item{plot}{whether to produce scree plot of RDA axes (defaults to TRUE)}
}
\value{
results from outlier tests. If \code{outlier_method = "p"}, a list of outlier snps, p-values, and results from rdadapt test (see Capblancq & Forester 2021; https://github.com/Capblancq/RDA-landscape-genomics/blob/main/RDA_landscape_genomics.Rmd). If \code{outlier_method = "z"}, a dataframe with outlier snp z-scores for each axes
}
\description{
Get significant outliers from RDA model
}
