% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_run}
\alias{rda_run}
\title{Run RDA}
\usage{
rda_run(
  gen,
  env,
  coords,
  model = "full",
  sig = 0.05,
  sig_method = "p",
  padj_method = "fdr",
  Pin = 0.05,
  R2permutations = 1000,
  R2scope = TRUE,
  lm_test = FALSE
)
}
\arguments{
\item{gen}{genotype matrix}

\item{env}{dataframe with environmental data}

\item{coords}{dataframe with coordinates (only needed if K selection is performed with TESS)}

\item{model}{whether to use all variables ("full") or perform variable selection ("best")}

\item{sig}{alpha level}

\item{padj_method}{correction method supplied to \code{p.adjust} (can be "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{R2permutations}{Number of permutations used in the estimation of
  adjusted \eqn{R^2}{R2} for \code{\link[vegan]{cca}} using
  \code{\link[vegan]{RsquareAdj}}.
}

\item{R2scope}{
  Use adjusted \eqn{R^2}{R2} as the stopping criterion: only models with
  lower adjusted \eqn{R^2}{R2} than scope are accepted.
  }
}
\description{
Run RDA
}
