% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_run}
\alias{rda_run}
\title{Run RDA}
\usage{
rda_run(
  gen,
  env,
  coords = NULL,
  model = "full",
  correctGEO = FALSE,
  correctPC = FALSE,
  nPC = 3,
  Pin = 0.05,
  R2permutations = 1000,
  R2scope = T
)
}
\arguments{
\item{gen}{genotype dosage matrix (rows = individuals & columns = SNPs) or \code{vcfR} object}

\item{env}{dataframe with environmental data or a Raster* type object from which environmental values for the coordinates can be extracted}

\item{coords}{dataframe with coordinates (only needed if correctGEO = TRUE) or if env is a Raster* from which values should be extracted}

\item{model}{whether to fit the model with all variables ("full") or to perform variable selection to determine the best set of variables ("best"); defaults to "full"}

\item{correctGEO}{whether to condition on geographic coordinates}

\item{correctPC}{whether to condition on PCs from PCA of genotypes}

\item{nPC}{number of PCs to use if correctPC = TRUE (defaults to 3); if set to "manual" a selection option with a terminal prompt will be provided}

\item{Pin}{if \code{model = "best"}, limits of permutation P-values for adding (\code{Pin}) a term to the model, or dropping (\code{Pout}) from the model. Term is added if\code{ P <= Pin}, and removed if \code{P > Pout} (see \link[vegan]{ordiR2step})}

\item{R2permutations}{if \code{model = "best"}, number of permutations used in the estimation of adjusted R2 for cca using RsquareAdj (see \link[vegan]{ordiR2step})}

\item{R2scope}{if \code{model = "best"}, use adjusted R2 as the stopping criterion: only models with lower adjusted R2 than scope are accepted (see \link[vegan]{ordiR2step})}
}
\value{
RDA model
}
\description{
Run RDA
}
\seealso{
Other RDA functions: 
\code{\link{rda_cor}()},
\code{\link{rda_do_everything}()},
\code{\link{rda_getoutliers}()},
\code{\link{rda_plot}()},
\code{\link{rda_table}()},
\code{\link{rda_varpart_table}()},
\code{\link{rda_varpart}()}
}
\concept{RDA functions}
