% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_run}
\alias{rda_run}
\title{Run RDA}
\usage{
rda_run(
  gen,
  env,
  coords = NULL,
  model = "full",
  correctGEO = FALSE,
  correctPC = FALSE,
  nPC = 3,
  Pin = 0.05,
  R2permutations = 1000,
  R2scope = T
)
}
\value{
RDA model
}
\description{
Run RDA
}
