% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_table}
\alias{rda_table}
\title{Make publication quality table of correlation test results}
\usage{
rda_table(
  cor_df,
  sig = 0.05,
  sig_only = TRUE,
  top = FALSE,
  order = FALSE,
  var = NULL,
  nrow = NULL,
  digits = 2
)
}
\arguments{
\item{cor_df}{dataframe of correlation results output from \link[algatr]{rda_cor}}
}
\value{
An object of class \code{gt_tbl}
}
\description{
Make publication quality table of correlation test results
}
