% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_table}
\alias{rda_table}
\title{Create \code{gt} table of RDA results}
\usage{
rda_table(
  cor_df,
  sig = 0.05,
  sig_only = TRUE,
  top = FALSE,
  order = FALSE,
  var = NULL,
  nrow = NULL,
  digits = 2
)
}
\arguments{
\item{cor_df}{dataframe of correlation results output from \link[algatr]{rda_cor}}

\item{sig}{if \code{outlier_method = "p"}, the significance level to use to identify SNPs (defaults to 0.05)}

\item{sig_only}{whether to only include loci with p-values less than \code{sig} (defaults to TRUE)}

\item{top}{whether to only include only keep the top variable for each snp in the table by the strength of the correlation (defaults to FALSE)}

\item{order}{whether to order by the magnitude of the correlation (defaults to FALSE)}

\item{var}{which variables to include (defaults to including all variables)}

\item{nrow}{number of rows to display (defaults to displaying all rows)}

\item{digits}{number of digits to include (defaults to 2)}
}
\value{
An object of class \code{gt_tbl}
}
\description{
Create \code{gt} table of RDA results
}
\seealso{
Other RDA functions: 
\code{\link{rda_cor}()},
\code{\link{rda_do_everything}()},
\code{\link{rda_getoutliers}()},
\code{\link{rda_plot}()},
\code{\link{rda_run}()},
\code{\link{rda_varpart_table}()},
\code{\link{rda_varpart}()}
}
\concept{RDA functions}
