% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_tidy}
\alias{rda_tidy}
\title{Make dataframe for ggplot from RDA results}
\usage{
rda_tidy(rda_mod, rda_loci, naxes)
}
\arguments{
\item{rda_mod}{rda model}

\item{rda_loci}{rda candidate loci}

\item{naxes}{number of axes to include}
}
\description{
Make dataframe for ggplot from RDA results
}
