% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_varpart}
\alias{rda_varpart}
\title{Partial RDA variance partitioning}
\usage{
rda_varpart(
  gen,
  env,
  coords,
  Pin,
  R2permutations,
  R2scope,
  nPC,
  digits = 2,
  call_col = FALSE
)
}
\arguments{
\item{digits}{number of digits to include (defaults to 2)}

\item{call_col}{whether to include column with RDA call (defaults to FALSE)}
}
\value{
df with relevant statistics from variance partitioning analysis
}
\description{
Partial RDA variance partitioning
}
