% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{rda_varpart}
\alias{rda_varpart}
\title{Partial RDA variance partitioning}
\usage{
rda_varpart(gen, env, coords, Pin, R2permutations, R2scope, nPC)
}
\arguments{
\item{gen}{genotype dosage matrix (rows = individuals & columns = SNPs) or \code{vcfR} object}

\item{env}{dataframe with environmental data or a Raster* type object from which environmental values for the coordinates can be extracted}

\item{coords}{dataframe with coordinates (only needed if correctGEO = TRUE) or if env is a Raster* from which values should be extracted}

\item{Pin}{if \code{model = "best"}, limits of permutation P-values for adding (\code{Pin}) a term to the model, or dropping (\code{Pout}) from the model. Term is added if\code{ P <= Pin}, and removed if \code{P > Pout} (see \link[vegan]{ordiR2step}) (defaults to 0.05)}

\item{R2permutations}{if \code{model = "best"}, number of permutations used in the estimation of adjusted R2 for cca using RsquareAdj (see \link[vegan]{ordiR2step}) (defaults to 1000)}

\item{R2scope}{if \code{model = "best"} and set to TRUE (default), use adjusted R2 as the stopping criterion: only models with lower adjusted R2 than scope are accepted (see \link[vegan]{ordiR2step})}

\item{nPC}{number of PCs to use if correctPC = TRUE (defaults to 3); if set to "manual" a selection option with a terminal prompt will be provided}
}
\value{
df with relevant statistics from variance partitioning analysis
}
\description{
Partial RDA variance partitioning
}
\seealso{
Other RDA functions: 
\code{\link{rda_cor}()},
\code{\link{rda_do_everything}()},
\code{\link{rda_getoutliers}()},
\code{\link{rda_plot}()},
\code{\link{rda_run}()},
\code{\link{rda_table}()},
\code{\link{rda_varpart_table}()}
}
\concept{RDA functions}
