% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{rm_islands}
\alias{rm_islands}
\title{Remove islands from mapping}
\usage{
rm_islands(input, shape, min_vertices = 10000)
}
\arguments{
\item{input}{raster (SpatRaster or Raster*) or coordinates (sf or two column dataframe) to mask or remove island points from}

\item{shape}{SpatialPolygons, sf, or sf object to create island mask}

\item{min_vertices}{minimum number of vertices in polygons to retain (defaults to 10000)}
}
\value{
object (of input type) with islands removed
}
\description{
Remove islands from mapping
}
