% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{rm_islands}
\alias{rm_islands}
\title{Remove islands from mapping
TODO: fill in param details}
\usage{
rm_islands(input, shape, min_vertices = 10000)
}
\arguments{
\item{min_vertices}{}
}
\description{
Remove islands from mapping
TODO: fill in param details
}
