% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{rm_islands}
\alias{rm_islands}
\title{Remove islands from mapping}
\usage{
rm_islands(input, shape, min_vertices = 10000)
}
\arguments{
\item{input}{RasterLayer or RasterStack object with islands to be removed; also accepts coords}

\item{shape}{spatial object to filter TODO \link{EAC}: needs more detail}

\item{min_vertices}{minimum number of vertices in polygons to retain (defaults to 10000)}
}
\description{
Remove islands from mapping
}
