% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDM.R
\name{scale01}
\alias{scale01}
\title{Scale genetic distances from 0 to 1}
\usage{
scale01(x)
}
\arguments{
\item{x}{genetic distance matrix}
}
\value{
genetic distance matrix scaled from 0 to 1
}
\description{
Scale genetic distances from 0 to 1
}
\seealso{
Other GDM functions: 
\code{\link{gdm_df}()},
\code{\link{gdm_do_everything}()},
\code{\link{gdm_map}()},
\code{\link{gdm_plot_isplines}()},
\code{\link{gdm_plot_vars}()},
\code{\link{gdm_run}()},
\code{\link{gdm_varimp}()}
}
\concept{GDM functions}
