% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{scaleRGB}
\alias{scaleRGB}
\title{Scale three layers of environmental data to R, G, and B for mapping}
\usage{
scaleRGB(env)
}
\arguments{
\item{env}{object of type SpatRaster or RasterStack}
}
\description{
Scale three layers of environmental data to R, G, and B for mapping
}
