% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{simple_impute}
\alias{simple_impute}
\title{Impute NA values
NOTE: use extreme caution when using this form of simplistic imputation. We mainly provide this code for creating test datasets and highly discourage its use in analyses.}
\usage{
simple_impute(x, FUN = median)
}
\arguments{
\item{x}{Matrix}

\item{f}{Function to use for imputation (defaults to median)}
}
\description{
Impute NA values
NOTE: use extreme caution when using this form of simplistic imputation. We mainly provide this code for creating test datasets and highly discourage its use in analyses.
}
