% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_do_everything}
\alias{tess_do_everything}
\title{TESS function to do everything}
\usage{
tess_do_everything(
  gen,
  coords,
  grid,
  Kvals = 1:10,
  K_selection = "auto",
  plot_method = "maxQ",
  col_breaks = 20,
  col_alpha = 0.5,
  minQ = 0.1,
  tess_method = "projected.ls",
  ploidy = 2,
  correct_kriged_Q = TRUE
)
}
\arguments{
\item{gen}{genotype dosage matrix (rows = individuals & columns = snps) or \code{vcfR} object}

\item{coords}{dataframe with x and y coordinates}

\item{grid}{RasterLayer or other gridded spatial object for kriging}

\item{Kvals}{vector of K values to test}

\item{K_selection}{how to perform K selection (options: "auto" for automatic selection based on \link[algatr]{bestK} (default) or "manual" to enter into console)}

\item{ploidy}{an integer corresponding to ploidy of the studied
organism. Haploids have ploidy = 1, diploids have ploidy = 2, etc.}

\item{correct_kriged_Q}{whether to correct kriged Q values so values greater than 1 are set to 1 and values less than 0 are set to 0 (defaults to TRUE)}
}
\value{
list with all TESS results, final K value, and final kriged raster
}
\description{
TESS function to do everything
}
\concept{TESS functions}
