% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_ggplot}
\alias{tess_ggplot}
\title{ggplot of TESS results}
\usage{
tess_ggplot(
  krig_admix,
  ggplot_fill = ggplot2::scale_fill_viridis_d(option = "turbo"),
  coords = NULL
)
}
\arguments{
\item{krig_admix}{RasterStack returned by \link[landgen]{tess_krig}}

\item{ggplot_fill}{any ggplot2 scale fill discrete function (default: \link[landgen]{scale_fill_viridis_d}, \code{option = "turbo"})}

\item{coords}{dataframe with x and y coordinates for plotting (optional)}
}
\description{
ggplot of TESS results
}
