% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_ggplot}
\alias{tess_ggplot}
\title{ggplot of TESS results}
\usage{
tess_ggplot(
  krig_admix,
  coords = NULL,
  plot_method = "maxQ",
  ggplot_fill = algatr_col_default("ggplot"),
  minQ = 0.1,
  plot_axes = FALSE,
  rel_widths = c(3, 1)
)
}
\arguments{
\item{krig_admix}{SpatRaster returned by \link[algatr]{tess_krig}}

\item{coords}{dataframe with x and y coordinates for plotting (optional)}

\item{plot_method}{method for making rainbow map of kriged layers (options: "maxQ" to only plot the max Q value for each cell (default), "allQ" to plot all Q values greater than \code{minQ}, "maxQ_poly" or "allQ_poly" to create the plots as previously described, but as polygons for each K instead of continuous Q values)}

\item{ggplot_fill}{any ggplot2 scale fill discrete function (default: \link[algatr]{scale_fill_viridis_d}, \code{option = "turbo"})}

\item{minQ}{threshold for minimum Q-value for rainbow plotting if \code{plot_method = "allQ"} or \code{plot_method = "allQ_poly"} is used (defaults to 0.10)}

\item{plot_axes}{whether to plot axes or not (defaults to FALSE)}

\item{rel_widths}{if \code{plot_method = "maxQ"} or \code{plot_method = "allQ"} is used, sets relative widths of kriged TESS map and legend (defaults to 3:1), from \link[cowplot]{plot_grid}}
}
\value{
ggplot object of TESS results
}
\description{
ggplot of TESS results
}
\seealso{
Other TESS functions: 
\code{\link{allK_plot_helper}()},
\code{\link{bestK}()},
\code{\link{ggbarplot_helper}()},
\code{\link{krig_legend}()},
\code{\link{pops_helper}()},
\code{\link{tess_barplot}()},
\code{\link{tess_col_default}()},
\code{\link{tess_do_everything}()},
\code{\link{tess_ggbarplot}()},
\code{\link{tess_krig}()},
\code{\link{tess_ktest}()},
\code{\link{tess_plot_allK}()}
}
\concept{TESS functions}
