% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_ggplot}
\alias{tess_ggplot}
\title{ggplot of TESS results}
\usage{
tess_ggplot(
  krig_admix,
  coords = NULL,
  plot_method = "maxQ",
  ggplot_fill = algatr_col_default("ggplot"),
  minQ = 0.1,
  plot_axes = FALSE
)
}
\arguments{
\item{krig_admix}{SpatRaster returned by \link[algatr]{tess_krig}}

\item{coords}{dataframe with x and y coordinates for plotting (optional)}

\item{plot_method}{method for making rainbow map of kriged layers (options: "maxQ" to only plot the max Q value for each cell (default), "allQ" to plot all Qvalues greater than \code{minQ}, "maxQ_poly" or "allQ_poly" to create the plots as previously described, but as polygons for each K instead of continuous Q values)}

\item{ggplot_fill}{any ggplot2 scale fill discrete function (default: \link[algatr]{scale_fill_viridis_d}, \code{option = "turbo"})}

\item{minQ}{threshold for minimum Q-value for rainbow plotting if \code{method = "all"} is used (defaults to 0.10)}
}
\description{
ggplot of TESS results
}
\seealso{
Other TESS functions: 
\code{\link{allK_plot_helper}()},
\code{\link{bestK}()},
\code{\link{tess_barplot}()},
\code{\link{tess_col_default}()},
\code{\link{tess_do_everything}()},
\code{\link{tess_krig}()},
\code{\link{tess_ktest}()},
\code{\link{tess_plot_allK}()},
\code{\link{tess_plot}()}
}
\concept{TESS functions}
