% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_krig}
\alias{tess_krig}
\title{Krige admixture values}
\usage{
tess_krig(qmat, coords, spdf, n_cell = 10000)
}
\arguments{
\item{qmat}{qmatrix}

\item{coords}{dataframe with x and y coordinates}

\item{spdf}{Spatial Points Dataframe to create grid for kriging}

\item{n_cell}{number of grid cells to use when kriging}
}
\description{
Krige admixture values
}
