% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_krig}
\alias{tess_krig}
\title{Krige admixture values}
\usage{
tess_krig(
  qmat,
  coords,
  grid = NULL,
  correct_kriged_Q = TRUE,
  krig_method = "autoKrige"
)
}
\arguments{
\item{qmat}{qmatrix}

\item{coords}{coordinates of samples as sf points, a two-column matrix, or a data.frame representing x and y coordinates (see Details for important information about projections)}

\item{grid}{SpatRaster for kriging}

\item{correct_kriged_Q}{whether to correct kriged Q values so values greater than 1 are set to 1 and values less than 0 are set to 0 (defaults to TRUE)}

\item{krig_method}{type of kriging to perform; options are "autoKrige" (default; uses \code{automap::autoKrige()} function) or "fields" (uses \code{fields::Krig()} function)}
}
\value{
Raster\* type object of kriged Q values
}
\description{
Krige admixture values
}
\seealso{
Other TESS functions: 
\code{\link{allK_plot_helper}()},
\code{\link{bestK}()},
\code{\link{ggbarplot_helper}()},
\code{\link{krig_legend}()},
\code{\link{pops_helper}()},
\code{\link{tess_barplot}()},
\code{\link{tess_col_default}()},
\code{\link{tess_do_everything}()},
\code{\link{tess_ggbarplot}()},
\code{\link{tess_ggplot}()},
\code{\link{tess_ktest}()},
\code{\link{tess_plot_allK}()}
}
\concept{TESS functions}
