% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_ktest}
\alias{tess_ktest}
\title{Test multiple K values}
\usage{
tess_ktest(
  gen,
  coords,
  Kvals = 1:10,
  grid = NULL,
  tess_method = "projected.ls",
  K_selection = "auto",
  ploidy = 2
)
}
\description{
Test multiple K values
}
