% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_ktest}
\alias{tess_ktest}
\title{Test multiple K values}
\usage{
tess_ktest(
  gen,
  coords,
  Kvals = 1:10,
  spdf = NULL,
  tess_method = "projected.ls",
  ploidy = 2,
  k_selection = "auto"
)
}
\arguments{
\item{gen}{genotype matrix}

\item{coords}{dataframe with x and y coordinates}

\item{Kvals}{vector of K values to test}

\item{tess_method}{method provided to tess}

\item{ploidy}{ploidy of organism (defaults to 2)}

\item{k_selection}{how to perform k selection (options: "auto" for automatic selection based on \code{\link{bestK}} or "manual" to enter into console)}
}
\description{
Test multiple K values
}
