% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_ktest}
\alias{tess_ktest}
\title{Test multiple K values}
\usage{
tess_ktest(
  gen,
  coords,
  Kvals = 1:10,
  grid = NULL,
  tess_method = "projected.ls",
  K_selection = "auto",
  ploidy = 2
)
}
\arguments{
\item{gen}{genotype matrix}

\item{coords}{dataframe with x and y coordinates}

\item{Kvals}{vector of K values to test}

\item{grid}{RasterLayer or other gridded spatial object for kriging}

\item{K_selection}{how to perform k selection (options: "auto" for automatic selection based on \link[landgen]{bestK} (default) or "manual" to enter into console)}

\item{ploidy}{an integer corresponding to ploidy of the studied
organism. Haploids have ploidy = 1, diploids have ploidy = 2, etc.}
}
\description{
Test multiple K values
}
