% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess_plot_allK}
\alias{tess_plot_allK}
\title{Plot all kriged Q values for each K}
\usage{
tess_plot_allK(
  krig_admix,
  coords = NULL,
  col_pal = viridis::turbo,
  col_breaks = 20
)
}
\arguments{
\item{krig_admix}{RasterStack returned by \link[landgen]{tess_krig}}

\item{coords}{dataframe with x and y coordinates for plotting (optional)}
}
\description{
Plot all kriged Q values for each K
}
