% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{unfold}
\alias{unfold}
\title{unfold converts the lower diagonal elements of a matrix into a vector}
\usage{
unfold(X, scale = TRUE)
}
\arguments{
\item{X}{is a distance matrix}

\item{scale}{if TRUE then matrices will be standardized (defaults to TRUE)}
}
\description{
unfold converts the lower diagonal elements of a matrix into a vector
}
\seealso{
Other MMRR functions: 
\code{\link{MMRR}()},
\code{\link{mmrr_df}()},
\code{\link{mmrr_do_everything}()},
\code{\link{mmrr_plot}()},
\code{\link{mmrr_run}()},
\code{\link{mmrr_table}()},
\code{\link{mmrr_var_sel}()}
}
\concept{MMRR functions}
