% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingen.R
\name{wingen_do_everything}
\alias{wingen_do_everything}
\title{wingen function to do everything  (preview and generate moving window maps, krige, and mask)}
\usage{
wingen_do_everything(
  preview = FALSE,
  lyr,
  coords,
  wdim = 3,
  fact = 0,
  sample_count = TRUE,
  min_n = 2,
  gen,
  stat = "pi",
  rarify = FALSE,
  kriged = FALSE,
  grd = NULL,
  index = 1,
  agg_grd = NULL,
  disagg_grd = NULL,
  agg_r = NULL,
  disagg_r = NULL,
  masked = FALSE,
  mask = NULL,
  bkg = NULL,
  plot_count = FALSE
)
}
\arguments{
\item{preview}{whether to produce preview of raster layer, window and focal cell size using \link[wingen]{preview_gd} (default = FALSE)}

\item{lyr}{SpatRaster or RasterLayer to slide the window across (see Details for important information about projections)}

\item{coords}{coordinates of samples as sf points, a two-column matrix, or a data.frame representing x and y coordinates (see Details for important information about projections)}

\item{wdim}{dimensions (height x width) of window; if only one value is provided, a square window is created (defaults to 3 x 3 window)}

\item{fact}{aggregation factor to apply to \code{lyr} (defaults to 0; \emph{note:} increasing this value reduces computational time)}

\item{sample_count}{whether to create plot of sample counts for each cell (defaults to TRUE)}

\item{min_n}{min number of samples to use in calculations (any focal cell with a window containing less than this number of samples will be assigned a value of NA)}

\item{gen}{genetic data either as an object of type vcf or a path to a vcf file (\emph{note:} order matters! The coordinate and genetic data should be in the same order; there are currently no checks for this)}

\item{stat}{genetic diversity statistic to calculate (can either be \code{"pi"} for nucleotide diversity (default), \code{"Ho"} for average observed heterozygosity across all sites, \code{"allelic_richness"} for average number of alleles across all sites, or \code{"biallelic_richness"} to get average allelic richness across all sites for a biallelic dataset (this option is faster than \code{"allelic_richness"}))}

\item{rarify}{if rarify = TRUE, rarefaction is performed (defaults to FALSE)}

\item{kriged}{whether to smooth out mapped values using kriging using \link[wingen]{krig_gd} (default = FALSE)}

\item{grd}{object to create grid for kriging; can be a SpatRaster, RasterLayer, SpatialPointsDataFrame, or a gridded object as defined by 'sp'. If undefined, will use \code{r} to create a grid.}

\item{index}{integer indices of layers in raster stack to krige (defaults to 1; i.e., the first layer)}

\item{agg_grd}{factor to use for aggregation of \code{grd}, if provided (this will decrease the resolution of the final kriged raster; defaults to NULL)}

\item{disagg_grd}{factor to use for disaggregation of \code{grd}, if provided (this will increase the resolution of the final kriged raster; defaults to NULL)}

\item{agg_r}{factor to use for aggregation of \code{r}, if provided (this will decrease the number of points used in the kriging model; defaults to NULL)}

\item{disagg_r}{factor to use for disaggregation, of \code{r} if provided (this will increase the number of points used in the kriging model; defaults to NULL)}

\item{masked}{whether to mask out areas outside region of interest using \link[wingen]{mask_gd} (default = FALSE)}

\item{plot_count}{if TRUE, whether to visualize sample counts using \link[wingen]{plot_count} (default = FALSE)}
}
\value{
RasterBrick object final raster
}
\description{
wingen function to do everything  (preview and generate moving window maps, krige, and mask)
}
\details{
When using wingen, please cite the original citation: Bishop, A.P., Chambers, E.A., Wang, I.J. (2023). Generating continuous maps of genetic diversity using moving windows. Methods Ecol. Evol. doi: https://doi.org/10.1111/2041-210X.14090
N.B.: Be aware that this function sets \emph{many} of the wingen function arguments to defaults, which my result in sub optimal results. We highly advise researchers to run each wingen function separately for best results.
}
\seealso{
Other wingen functions: 
\code{\link{krig_agg_helper}()},
\code{\link{krig_helper}()}
}
\concept{wingen functions}
